### 2025-02-04 #7746 @editorialbot
[PRE REVIEW]: Sorcha: A Solar System Survey Simulator for the Legacy Survey of Space and Time
**Submitting author:** @mschwamb (<a href="http://orcid.org/0000-0003-4365-1455">Megan E. Schwamb</a>)
**Repository:** https://github.com/dirac-institute/sorcha/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0 (the version in the github repo is v1.0 we are waiting to push to v1.0 after JOSS review with the publication of the paper)
