### 2025-02-04 #7746 @warrickball
Hi @mschwamb, thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!

In the meantime, can I double check that the code is dual licensed as BSD 3-Clause and GPL v3? Is this because some files fall under one licence and other files under the other? I don't think dual licensing is a problem, as long as both licences are open source (according to the [OSI definition](https://opensource.org/osd)), but we would like to know why you've chosen multiple licences.

I'm also not sure why the paper is failing to build. It works for me locally. Jeremy Kubica's affiliations look fine and there aren't any stray tabs. Any ideas, @openjournals/dev?
