### 2025-02-04 #7746 @mschwamb
`In the meantime, can I double check that the code is dual licensed as BSD 3-Clause and GPL v3? Is this because some files fall under one licence and other files under the other? I don't think dual licensing is a problem, as long as both licences are open source (according to the [OSI definition](https://opensource.org/osd)), but we would like to know why you've chosen multiple licences.`

Yes -we  license all contributions to this code under both BSDv3 and GPLv3 - We've heavily adapted some of the functions from the Rubin Observatory LSST stack (it is identified in the relevant files) and those were licensed originally under the GPLv3 so we've licensed under for both these reasons.
