### 2025-02-11 #7746 @warrickball
> Yes -we license all contributions to this code under both BSDv3 and GPLv3 - We've heavily adapted some of the functions from the Rubin Observatory LSST stack (it is identified in the relevant files) and those were licensed originally under the GPLv3 so we've licensed under for both these reasons.

Sorry for taking a few days to get back to you. I've been discussing this case and I think there's a licensing issue that needs to be resolved. Note that I am *not* a lawyer and none of this constitutes legal advice, but is my best interpretation of the situation based on my experience with open source licenses and after taking some advice from others on the editorial board.

The GPLv3-licensed files appear to me to be
```
src/sorcha/modules/PPAddUncertainties.py
src/sorcha/modules/PPDetectionProbability.py
src/sorcha/modules/PPFootprintFilter.py
src/sorcha/modules/PPRandomizeMeasurements.py
src/sorcha/modules/PPTrailingLoss.py
```
which all have the GPLv3 notice in them. Note that, since the authors of Sorcha are presumably not the original copyright holders, these cannot simply be relicensed as GPLv3 & BSD 3-Clause, because GPLv3's copyleft terms require that the code is shared under licences that preserve the copyleft terms of GPLv3.

For the remaining code, presuming the Sorcha authors wrote it all, you're free to license under GPLv3 & BSD 3-Clause, although I note this is an unusual choice because BSD 3-Clause is a permissive license anyway. This gets murky but dual-licensing under these two license does have merit for e.g. inclusion in other GPLv3 code, where everything could be under one licence (GPLv3).

If you stick to licensing the code you've written as GPLv3 & BSD 3-Clause, then we need a very clear statement somewhere—`README.md` or `LICENSE.txt` are options—specifying which files are under which licence. Alternatively, if Sorcha is licensed as GPLv3 only, then the whole project would fall under that single licence.

I'm happy to discuss further, since it looks (based on the last change to `LICENSE.txt`) that the Sorcha team may have already discussed this in some detail.
