### 2025-02-12 #7746 @mschwamb
Mario Jurić on our team is the one who made the commit and has strong opinions on the licenses.  So I'll point him to this discussion, and we'll let you know what change we've made to address the issue.
