### 2025-02-12 b8940b7
Change ang_fov to ang_fov + buffer
