+++ src/sorcha/readers/EphemerisReader.py
@@ -1,4 +1,3 @@
-import numpy as np
 import logging
 import sys
 
