+++ tests/ephemeris/test_simulation_driver.py
@@ -1,4 +1,3 @@
-import pytest
 import numpy as np
 import pandas as pd
 from sorcha.ephemeris.simulation_driver import calculate_rates_and_geometry, EphemerisGeometryParameters
