+++ src/sorcha/ephemeris/simulation_driver.py
@@ -15,9 +15,8 @@ from sorcha.ephemeris.simulation_setup import (
 from sorcha.ephemeris.simulation_constants import *
 from sorcha.ephemeris.simulation_geometry import *
 from sorcha.ephemeris.simulation_parsing import *
-from sorcha.utilities.dataUtilitiesForTests import get_data_out_filepath
 from sorcha.ephemeris.pixel_dict import PixelDict
-from sorcha.modules.PPOutput import PPOutWriteCSV, PPOutWriteSqlite3, PPOutWriteHDF5
+from sorcha.modules.PPOutput import PPOutWriteCSV, PPOutWriteHDF5
 
 
 @dataclass
@@ -35,11 +34,13 @@ class EphemerisGeometryParameters:
 
 def get_vec(row, vecname):
     """
-    Extracts a vector from a Pandas dataframe row
+    Extracts a vector from a Pandas dataframe row.
+
     Parameters
     ----------
     row : row from the dataframe
     vecname : name of the vector
+
     Returns
     -------
     : 3D numpy array
@@ -123,7 +124,6 @@ def create_ephemeris(orbits_df, pointings_df, args, sconfigs):
     furnish_spiceypy(args, sconfigs.auxiliary)
     verboselog("Generating ASSIST+REBOUND simulations.")
     sim_dict = generate_simulations(ephem, gm_sun, gm_total, orbits_df, args)
-    pixel_dict = defaultdict(list)
     observatories = Observatory(args, sconfigs.auxiliary)
 
     output = StringIO()
@@ -244,17 +244,19 @@ def get_residual_vectors(v1):
     Decomposes the vector into two unit vectors to facilitate computation of on-sky angles
     The decomposition is such that A  = (-sin (RA), cos(RA), 0) is in the direction of increasing RA,
     and D = (-sin(dec)cos (RA), -sin(dec) sin(RA), cos(dec)) is in the direction of increasing Dec
-    The triplet (A,D,v1) forms an orthonormal basis of the 3D vector space
+    The triplet (A,D,v1) forms an orthonormal basis of the 3D vector space.
+
     Parameters
-    -----------
-        v1 : array, shape = (3,))
-            The vector to be decomposed
-    Returns
     ----------
-        A :  array, shape = (3,))
-            A  vector
-        D : array, shape = (3,))
-            D vector
+    v1 : array, shape = (3,))
+        The vector to be decomposed
+
+    Returns
+    -------
+    A :  array, shape = (3,))
+        A  vector
+    D : array, shape = (3,))
+        D vector
     """
     x, y, z = v1
     cosd = np.sqrt(1 - z * z)
