### 2025-02-16 68d96f6
Update LICENSE.txt
clarify licenses
