### 2025-02-16 #7746 @mschwamb
@warrickball I spoke to Mario about the licensing. What you stated is what intended. I'm still learning my way around licensing so I did not communicate it very clarify. We agree that we need to clearly state which files are under which license in the license file. I've done that in this [PR](https://github.com/dirac-institute/sorcha/pull/1130) that should be merged sometime on Monday. Please let me know if JOSS prefers a different wording or more information in the license file.
