### 2025-02-17 #7746 @mschwamb
The license update has now been merged into the main branch.
