### 2025-02-18 a265414
Made a separated function for DESFadingFunctionFilter
Made separate files for DES's fading function and detection probability and removed the changes I made to the original functions
