+++ src/sorcha/modules/DESDetectionProbability.py
@@ -20,11 +20,11 @@
 import numpy as np
 
 
-def calcDetectionProbability(mag, limmag, fillFactor=1.0, w=0.1):
+def DEScalcDetectionProbability(mag, limmag, c, k):
     """
     Find the probability of a detection given a visual magnitude,
-    limiting magnitude, and fill factor, determined by the fading function
-    from Veres & Chesley (2017).
+    limiting magnitude, a scaling factor c, and transition sharpness k. Equation from
+    Bernardinelli et al., 2022
 
     Parameters
     -----------
@@ -34,11 +34,11 @@ def calcDetectionProbability(mag, limmag, fillFactor=1.0, w=0.1):
     limmag : float or array of floats
         Limiting magnitude of the field.
 
-    fillFactor : float), optional
-        Fraction of FOV covered by the camera sensor. Default = 1.0
+    c : float or array of floats
+        scaling factor
 
-    w : float
-        Distribution parameter. Default = 0.1
+    k : float or array of floats
+        transition sharpness
 
     Returns
     ----------
@@ -46,20 +46,18 @@ def calcDetectionProbability(mag, limmag, fillFactor=1.0, w=0.1):
         Probability of detection.
     """
 
-    P = fillFactor / (1.0 + np.exp((mag - limmag) / w))
+    P = c / (1 + np.exp(k * (mag - limmag)))
 
     return P
 
 
-def PPDetectionProbability(
+def DESDetectionProbability(
     eph_df,
-    trailing_losses=False,
-    trailing_loss_name="dmagDetect",
     magnitude_name="PSFMag",
     limiting_magnitude_name="fiveSigmaDepth_mag",
-    field_id_name="FieldID",
-    fillFactor=1.0,
-    w=0.1,
+    scaling_factor_name = "c",
+    transition_sharpness_name = "k"
+
 ):
     """
     Find probability of observations being observable for objectInField output.
@@ -71,11 +69,6 @@ def PPDetectionProbability(
     eph_df : Pandas dataframe
         Dataframe of observations.
 
-    trailing_losses : Boolean, optional
-        Are trailing losses being applied?, Default = False
-
-    trailing_loss_name : string, optional
-        eph_df column name for trailing losses, Default = dmagDetect
 
     magnitude_name : string, optional
         eph_df column name for observation limiting magnitude
@@ -89,11 +82,13 @@ def PPDetectionProbability(
         eph_df column name for observation field_id
         Default = FieldID
 
-    fillFactor : float, optional
-        Fraction of FOV covered by the camera sensor. Default = 1.0
+    scaling_factor_name: str, optional
+        eph_df column name for scaling factor
+        Default: c
 
-    w : float
-        Distribution parameter. Default =0.1
+    transition_sharpness_name: str, optional
+        eph_df column name for transition_sharpness
+        DEfault: k
 
     Returns
     ----------
@@ -102,14 +97,6 @@ def PPDetectionProbability(
 
     """
 
-    if not trailing_losses:
-        return calcDetectionProbability(
-            eph_df[magnitude_name], eph_df[limiting_magnitude_name], fillFactor, w
-        )
-    elif trailing_losses:
-        return calcDetectionProbability(
-            eph_df[magnitude_name] + eph_df[trailing_loss_name],
-            eph_df[limiting_magnitude_name],
-            fillFactor,
-            w,
-        )
+    return DEScalcDetectionProbability(
+        eph_df[magnitude_name], eph_df[limiting_magnitude_name], eph_df[scaling_factor_name], eph_df[transition_sharpness_name]
+    )
