### 2025-02-20 6c4f958
cleaned up DES fading function and added a camera footprint for DES
Cleaned up parts of the DES fading function code. Added a camera footprint for DES that uses the shapely package to find possible detections.
