+++ src/sorcha/modules/DESFootprintFilter.py
@@ -1,0 +1,392 @@
+# Developed for the Vera C. Rubin Observatory/LSST Data Management System.
+# This product includes software developed by the
+# Vera C. Rubin Observatory/LSST Project (https://www.lsst.org).
+#
+# Copyright 2020 University of Washington
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+import logging
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+import sys
+import importlib_resources
+from shapely.geometry import Point, MultiPolygon, Polygon
+
+from numba import njit
+
+deg2rad = np.radians
+sin = np.sin
+cos = np.cos
+
+logger = logging.getLogger(__name__)
+
+
+# ==============================================================================
+# coordinate transforms
+# ==============================================================================
+
+
+def radec_to_tangent_plane(ra, dec, field_ra, field_dec):
+    """
+    Converts ra and dec to xy on the plane tangent to image center, in the 2-d coordinate system where y is aligned with the meridian.
+
+    Parameters:
+    -----------
+    ra (float/array of floats): observation Right Ascension, radians.
+
+    dec (float/array of floats): observation Declination, radians.
+
+    fieldra (float/array of floats): field pointing Right Ascension, radians.
+
+    fielddec (float/array of floats): field pointing Declination, radians.
+
+    fieldID (float/array of floats): Field ID, optional.
+
+    Returns:
+    ----------
+    x, y (float/array of floats): Coordinates on the focal plane, radians projected
+    to the plane tangent to the unit sphere.
+
+    """
+
+    # convert to cartesian coordiantes on unit sphere
+    observation_vectors = np.array([cos(ra) * np.cos(dec), sin(ra) * np.cos(dec), sin(dec)])  # x  # y  # z
+
+    field_vectors = np.array(
+        [cos(field_ra) * np.cos(field_dec), sin(field_ra) * np.cos(field_dec), sin(field_dec)]
+    )
+
+    # make the basis vectors for the fields of view
+    # the "x" basis is easy, 90 d rotation of the x, y components
+    focalx = np.zeros(field_vectors.shape)
+    focalx[0] = -field_vectors[1]
+    focalx[1] = field_vectors[0]
+
+    # "y" by taking cross product of field vector and "x"
+    focaly = np.cross(field_vectors, focalx, axis=0)
+
+    # normalize
+    focalx /= np.linalg.norm(focalx, axis=0)
+    focaly /= np.linalg.norm(focaly, axis=0)
+
+    # extend observation vectors to plane tangent to field pointings
+    k = 1.0 / np.sum(field_vectors * observation_vectors, axis=0)
+    observation_vectors *= k
+    observation_vectors -= field_vectors
+
+    # get observation vectors as combinations of focal vectors
+    x = np.sum(observation_vectors * focalx, axis=0)
+    y = np.sum(observation_vectors * focaly, axis=0)
+
+    return x, y
+
+
+def radec_to_focal_plane(ra, dec, field_ra, field_dec, field_rot):
[…]

[…]
