### 2025-02-25 9587bdc
Added boostrap -f troubleshooting trick
