### 2025-03-05 #7746 @mschwamb
We've gotten our AAS journals review back and it's minor corrections. The referee has declined to see the paper again. So once the editor reviews our changes, we expect acceptance of the paper. Is there anything that can be done to move the JOSS review forward? I appreciate there's a back long and editing/peer review is a voluntary process so that there might not be anything that can be done to expedite this process.
