### 2025-03-12 f0a8257
Update paper.bib
updated journal names to note use abbreviations
