+++ JOSS_submission/paper.bib
@@ -2,7 +2,7 @@
 @ARTICLE{acton1996,
     author = {{Acton}, Charles H.},
     title = {Ancillary data services of  NASA's Navigation and Ancillary Information Facility},
-    journal = {\planss},
+    journal = {Planet.~Space~Sci},
     year = 1996,
     month = jan,
     volume = {44},
@@ -16,7 +16,7 @@
 @ARTICLE{acton2018,
     author = {{Acton}, Charles and {Bachman}, Nathaniel and {Semenov}, Boris and {Wright}, Edward},
     title = "{A look towards the future in the handling of space science mission geometry}",
-    journal = {\planss},
+    journal = {Planet.~Space~Sci},
     year = 2018,
     month = jan,
     volume = {150},
@@ -34,7 +34,7 @@
     Doi = {10.1051/0004-6361/201322068},
     Eid = {A33},
     Eprint = {1307.6212},
-    Journal = {\aap},
+    Journal = {A\&A},
     Keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
     Month = oct,
     Pages = {A33},
@@ -90,7 +90,7 @@
          {Weaver}, B.~A. and {Whitmore}, J.~B. and {Woillez}, J. and
          {Zabalza}, V. and {Astropy Contributors}},
         title = "{The Astropy Project: Building an Open-science Project and Status of the v2.0 Core Package}",
-      journal = {\aj},
+      journal = {AJ},
      keywords = {methods: data analysis, methods: miscellaneous, methods: statistical, reference systems, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2018,
         month = sep,
@@ -109,7 +109,7 @@ archivePrefix = {arXiv},
 @ARTICLE{astropy2022,
[…]
