### 2025-03-12 7e47d8c
Update paper.bib
spell out all the full name of the journals
