### 2025-03-12 b4ec48a
fix PSJ citation in paper.bib
PSJj -> Planetary Science Journal in hoover2022
