+++ JOSS_submission/paper.bib
@@ -268,7 +268,7 @@ archivePrefix = {arXiv},
 @ARTICLE{hoover2022,
        author = {{Hoover}, Devin J. and {Seligman}, Darryl Z. and {Payne}, Matthew J.},
         title = "{The Population of Interstellar Objects Detectable with the LSST and Accessible for In Situ Rendezvous with Various Mission Designs}",
-      journal = {PSJj},
+      journal = {Planetary Science Journal},
      keywords = {Interstellar objects, 52, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2022,
         month = mar,
