+++ src/sorcha/modules/DESFootprintFilter.py
@@ -20,10 +20,17 @@
 import logging
 import numpy as np
 import pandas as pd
-import matplotlib.pyplot as plt
+
 import sys
 import importlib_resources
-from shapely.geometry import Point, MultiPolygon, Polygon
+from shapely.geometry import Point, Polygon, MultiPolygon
+from shapely import STRtree, wkb
+import sqlite3
+import geopandas as gpd
+import matplotlib.pyplot as plt
+from scipy.spatial import KDTree
+from sklearn.neighbors import KDTree as skKD
+from tqdm import tqdm
 
 from numba import njit
 
@@ -133,8 +140,8 @@ class DESFootprint:
         """
 
         # file should be a .csv (and should be actually comma seperated)
-        # First column is the number of each detector
-        # other columns are strings of tuples of the points (I know its dumb gonna change it later)
+        # First column is the name of each detector
+        # the other two columns are the x, y postion of the detector corner
         # if the user doesn't provide their own version of the footprint,
         # we'll use the default DES version that comes included.
         if path:
@@ -177,7 +184,7 @@ class DESFootprint:
         polys = []
         poly_name_list = []
 
-        ccd_names = [
+        self.ccd_names = [
             "N1",
             "N2",
             "N3",
@@ -237,14 +244,11 @@ class DESFootprint:
             "S27",
             "S28",
             "S29",
-            "S30",
             "S31",
         ]
 
         # creates a polygon of each ccd using their corners
-        for detector in ccd_names:
-            if detector == "S30":
-                continue
+        for detector in self.ccd_names:
 
             corners = ccd_corners[ccd_corners["detectorId"] == detector].reset_index(drop=True)
 
@@ -260,17 +264,11 @@ class DESFootprint:
             polys.append(poly)
             poly_name_list.append(detector)
         # combines all polygons into a MultiPolygon
-        self.footprint = MultiPolygon(polys)
+        self.footprint = polys
         self.detectorId_list = poly_name_list
 
-    def applyDESFootprint(
-        self,
-        field_df,
-        ra_name="RA_deg",
-        dec_name="Dec_deg",
-        ra_name_field="fieldRA_deg",
-        dec_name_field="fieldDec_deg",
-        edge_thresh=None,
+    def applyDESFootprint_loadpolygons(
+        self, field_df, ra_name="RA_deg", dec_name="Dec_deg", fieldId="FieldID", edge_thresh=None
     ):
         """
         Determine whether detections fall on the sensors defined by the
@@ -308,85 +306,355 @@ class DESFootprint:
         detected : array
             Indices of rows in field_df which fall on the sensor(s).
 
+        detectorID :array
+            name of the detector that the object falls on.
         """
 
-        # convert detections to xyz on unit sphere
-        ra = deg2rad(field_df[ra_name])
-        dec = deg2rad(field_df[dec_name])
+        # connecting to visitors database and creating a list of polygons
+        with sqlite3.connect("/Users/ryanlyttle/Documents/Workstation/Dark energy survey /visits.db") as conn:
+            cursor = conn.cursor()
+            query = "SELECT geom FROM polygons_table WHERE observationId = ?"
+            index_ = set()
+            field_df_no_duplicates = field_df.drop_duplicates(subset=[fieldId], keep="first")
+            for obs_id in tqdm(field_df_no_duplicates[fieldId], desc="Processing Observations"):
 
-        # convert field pointings to xyz on unit sphere
-        fieldra = deg2rad(field_df[ra_name_field])
-        fielddec = deg2rad(field_df[dec_name_field])
+                row = cursor.execute(query, (obs_id,)).fetchone()
 
[…]
