+++ src/sorcha/data/DES/DES_config_file.ini
@@ -36,7 +36,7 @@ pointing_sql_query = SELECT observationId, observationMidpointMJD as observation
 # Configuration for running the ASSIST+REBOUND ephemerides generator.
 
 # the field of view of our search field, in degrees
-ar_ang_fov = 1.12
+ar_ang_fov = 1.3
 
 # the buffer zone around the field of view we want to include, in degrees
 ar_fov_buffer = 0.2
