+++ src/sorcha/ephemeris/__init__.py
@@ -30,5 +30,5 @@ from .simulation_driver import create_ephemeris
 
 from .orbit_conversion_utilities import (
     universal_cartesian,
-    universal_keplerian,
+    universal_cometary,
 )
