### 2025-03-13 #7746 @mschwamb
PDF regenerated to fix the journal titles
