### 2025-03-14 0e5022c
Wrap `create_assist_ephemeris` function so argparse object is not required.
