### 2025-03-14 94d4b47
changed DES config file and footprint filter
