+++ src/sorcha/modules/DESFootprintFilter.py
@@ -407,10 +407,14 @@ class DESFootprint:
 
             # attampt using sklearn Kdtree
             ccd_index = skKD(list(zip([row[8] for row in rows], [row[9] for row in rows])))
-            max_diff = max(abs(row[0] - row[8]) for row in rows)
-            polygon_index_to_create = ccd_index.query_radius(
-                points_query, max_diff, count_only=False, return_distance=False
+            # max_diff = max(abs(row[0] - row[8]) for row in rows)
+            # polygon_index_to_create = ccd_index.query_radius(
+            #     points_query, max_diff, count_only=False, return_distance=False
+            # )
+            polygon_index_to_create = ccd_index.query(
+                points_query, k=3,return_distance = False
             )
+            print(polygon_index_to_create)
             polygon_index_to_create = list(set(np.concatenate(polygon_index_to_create)))
             if not polygon_index_to_create:
                 continue
@@ -553,108 +557,97 @@ class DESFootprint:
 
         return detected, detectorID
 
-
-
     def applyDESFootprint_KDpolygons(
-            self, field_df, ra_name="RA_deg", dec_name="Dec_deg", fieldId="FieldID", edge_thresh=None
-        ):
-            """
-            Determine whether detections fall on the sensors defined by the
-            footprint. Also returns the an ID for the sensor a detection is made
-            on.
+        self, field_df, ra_name="RA_deg", dec_name="Dec_deg", fieldId="FieldID", edge_thresh=None
+    ):
+        """
+        Determine whether detections fall on the sensors defined by the
+        footprint. Also returns the an ID for the sensor a detection is made
+        on.
 
-            Parameters
-            -----------
-            field_df : Pandas dataframe
-                Dataframe containing detection information with pointings.
+        Parameters
+        -----------
+        field_df : Pandas dataframe
+            Dataframe containing detection information with pointings.
 
-            ra_name : string, optional
-                "field_df" dataframe's column name for object's RA
-                for the given observation. Default = "RA_deg" [units: degrees]
+        ra_name : string, optional
+            "field_df" dataframe's column name for object's RA
+            for the given observation. Default = "RA_deg" [units: degrees]
 
-            dec_name : string, optional
-                "field_df" dataframe's column name for object's declination
-                for the given observation. Default = "Dec_deg" [units: dgrees]
+        dec_name : string, optional
+            "field_df" dataframe's column name for object's declination
+            for the given observation. Default = "Dec_deg" [units: dgrees]
 
-            ra_name_field : string, optional
-                "field_df" dataframe's column name for the observation field's RA
-                Default = "fieldRA_deg" [units: degrees]
+        ra_name_field : string, optional
+            "field_df" dataframe's column name for the observation field's RA
+            Default = "fieldRA_deg" [units: degrees]
 
-            dec_name_field : string, optional
-                "field_df" dataframe's column name for the observation field's declination
-                Default = "fieldDec_deg" [Units: degrees]
-
-
-            edge_thresh: float, optional
-                An angular threshold in arcseconds for dropping pixels too close to the edge.
-                Default  = None
-
-            Returns
-            ----------
-            detected : array
-                Indices of rows in field_df which fall on the sensor(s).
-
-            detectorID :array
-                name of the detector that the object falls on.
-            """
-
-            detected_index = set()
-
-            # SQLite Database Connection
-            with sqlite3.connect(
-                "/Users/ryanlyttle/Documents/Workstation/Dark energy survey /visits_cent.db"
-            ) as conn:
-                cursor = conn.cursor()
-                query = "SELECT ra1,dec1,ra2,dec2,ra3,dec3,ra4,dec4,racenter,deccenter FROM observations WHERE observationId = ?"
[…]
