+++ src/sorcha/data/DES/DES_config_file.ini
@@ -110,34 +110,34 @@ fading_function_on = True
 
 [LINKINGFILTER]
 
-; # SSP detection efficiency. Which fraction of the objects will
-; # the automated Rubin Solar System Processing (SSP) pipeline successfully link? Float.
-; SSP_detection_efficiency = 0.95
+# SSP detection efficiency. Which fraction of the objects will
+# the automated Rubin Solar System Processing (SSP) pipeline successfully link? Float.
+SSP_detection_efficiency = 1
 
 
-; # Length of tracklets. How many observations of an object during one night are
-; # required to produce a valid tracklet?
-; SSP_number_observations = 2
+# Length of tracklets. How many observations of an object during one night are
+# required to produce a valid tracklet?
+SSP_number_observations = 1
 
-; # Minimum separation (in arcsec) between two observations of an object required 
-; # for the linking software to distinguish them as separate and therefore as a valid tracklet.
-; SSP_separation_threshold = 0.5
+# Minimum separation (in arcsec) between two observations of an object required 
+# for the linking software to distinguish them as separate and therefore as a valid tracklet.
+SSP_separation_threshold = 0.0002
 
-; # Maximum time separation (in days) between subsequent observations in a tracklet. 
-; # Default is 0.0625 days (90mins).
-; SSP_maximum_time = 0.0625
+# Maximum time separation (in days) between subsequent observations in a tracklet. 
+# Default is 0.0625 days (90mins).
+SSP_maximum_time = 0.5
 
-; # Number of tracklets for detection. How many tracklets are required to classify
-; # an object as detected? 
-; SSP_number_tracklets = 3
+# Number of tracklets for detection. How many tracklets are required to classify
+# an object as detected? 
+SSP_number_tracklets = 7
 
-; # The number of tracklets defined above must occur in <= this number of days to 
-; # constitute a complete track/detection.
-; SSP_track_window = 15
+# The number of tracklets defined above must occur in <= this number of days to 
+# constitute a complete track/detection.
+SSP_track_window = 2192
 
-; # The time in UTC at which it is noon at the observatory location (in standard time).
-; # For the LSST, 12pm Chile Standard Time is 4pm UTC.
-; SSP_night_start_utc = 16.0
+# The time in UTC at which it is noon at the observatory location (in standard time).
+# For the LSST, 12pm Chile Standard Time is 4pm UTC.
+SSP_night_start_utc = 19.0
 
 
 [OUTPUT]
