### 2025-03-19 #7746 @warrickball
> Is there anything that can be done to move the JOSS review forward?

In principle no but I don't like to keep AAS Journals waiting too long. Let me know when the paper is accepted (perhaps it already is?) and I'll try to nudge this up. Unfortunately yes, our backlog is currently a few months long.
