### 2025-03-24 b9b2979
added detailed summary for comet class
