### 2025-03-24 fda09c5
Cleaned up DESFootprintFilter.py
Cleaned up the DES fov filter and added a unit test for an error message in sorchaConfigs
