+++ src/sorcha/modules/DESFootprintFilter.py
@@ -24,94 +24,15 @@ import pandas as pd
 import sys
 import importlib_resources
 from shapely.geometry import Point, Polygon, MultiPolygon
-from shapely import STRtree, wkb
 import sqlite3
-import geopandas as gpd
-import matplotlib.pyplot as plt
 from scipy.spatial import KDTree
-from sklearn.neighbors import KDTree as skKD
-from tqdm import tqdm
 
-from numba import njit
 
 deg2rad = np.radians
 sin = np.sin
 cos = np.cos
 
 logger = logging.getLogger(__name__)
-
-
-# ==============================================================================
-# coordinate transforms
-# ==============================================================================
-
-
-def radec_to_tangent_plane(ra, dec, field_ra, field_dec):
-    """
-    Converts ra and dec to xy on the plane tangent to image center, in the 2-d coordinate system where y is aligned with the meridian.
-
-    Parameters:
-    -----------
-    ra (float/array of floats): observation Right Ascension, radians.
-
-    dec (float/array of floats): observation Declination, radians.
-
-    fieldra (float/array of floats): field pointing Right Ascension, radians.
-
-    fielddec (float/array of floats): field pointing Declination, radians.
-
-    fieldID (float/array of floats): Field ID, optional.
-
-    Returns:
-    ----------
-    x, y (float/array of floats): Coordinates on the focal plane, radians projected
-    to the plane tangent to the unit sphere.
-
-    """
-
-    # convert to cartesian coordiantes on unit sphere
-    observation_vectors = np.array([cos(ra) * np.cos(dec), sin(ra) * np.cos(dec), sin(dec)])  # x  # y  # z
-
-    field_vectors = np.array(
-        [cos(field_ra) * np.cos(field_dec), sin(field_ra) * np.cos(field_dec), sin(field_dec)]
-    )
-
-    # make the basis vectors for the fields of view
-    # the "x" basis is easy, 90 d rotation of the x, y components
-    focalx = np.zeros(field_vectors.shape)
-    focalx[0] = -field_vectors[1]
-    focalx[1] = field_vectors[0]
-
-    # "y" by taking cross product of field vector and "x"
-    focaly = np.cross(field_vectors, focalx, axis=0)
-
-    # normalize
-    focalx /= np.linalg.norm(focalx, axis=0)
-    focaly /= np.linalg.norm(focaly, axis=0)
-
-    # extend observation vectors to plane tangent to field pointings
-    k = 1.0 / np.sum(field_vectors * observation_vectors, axis=0)
-    observation_vectors *= k
-    observation_vectors -= field_vectors
-
-    # get observation vectors as combinations of focal vectors
-    x = np.sum(observation_vectors * focalx, axis=0)
-    y = np.sum(observation_vectors * focaly, axis=0)
-
-    return x, y
-
-
-def radec_to_focal_plane(ra, dec, field_ra, field_dec, field_rot):
-    # convert ra, dec to points on focal plane, x pointing to celestial north
-    x, y = radec_to_tangent_plane(ra, dec, field_ra, field_dec)
-    # rotate focal plane to align with detectors
-    xy = x + 1.0j * y
-    xy *= np.exp(1.0j * field_rot)  # which direction to rotate?
-
-    x = np.real(xy)
-    y = np.imag(xy)
-
-    return x, y
 
 
 # ==============================================================================
@@ -138,208 +59,13 @@ class DESFootprint:
             Footprint object for the provided sensors.
 
         """
+        # currently hard coded in.
+        # next step would be either to make them config.fov.attributes and cmdline args
 
-        # file should be a .csv (and should be actually comma seperated)
-        # First column is the name of each detector
-        # the other two columns are the x, y postion of the detector corner
[…]
