+++ src/sorcha/modules/DESApplyFOVFilter.py
@@ -45,16 +45,9 @@ def DESApplyFOVFilter(observations, sconfigs, module_rngs, footprint=None, verbo
     if sconfigs.fov.camera_model == "footprint":
         verboselog("Applying sensor footprint filter...")
 
-        onSensor, detectorIDs = footprint.applyDESFootprint_KDpolygons(
-            observations, edge_thresh=sconfigs.fov.footprint_edge_threshold
-        )
-        # onSensor, detectorIDs = footprint.applyDESFootprint_loadpolygons(
-        #     observations, edge_thresh=sconfigs.fov.footprint_edge_threshold
-        # )
+        onSensor = footprint.applyDESFootprint(observations)
 
         observations = observations.iloc[onSensor].copy()
-
-        # observations["detectorID"] = detectorIDs
 
         observations = observations.sort_index()
 
