### 2025-03-24 5c80e64
Added a new cmd line arg in sorcha run and a fovConfigs attritbute
Added a new cmd line arg in sorcha run for the visits database file. --vd, --visits_database

Added new fovConfigs attritbute. visits_query to query visits sql database.

now DESFootprintFilter is no longer hard coded and can be changed by the user.

also updated and added unit tests for these new changes
