+++ tests/sorcha/test_sorchaArguments.py
@@ -14,6 +14,7 @@ cmd_args_dict = {
     "loglevel": False,
     "pointing_database": get_demo_filepath("baseline_v2.0_1yr.db"),
     "stats": "./test.csv",
+    "visits_database": None,
 }
 
 
