+++ src/sorcha/data/DES/DES_config_file.ini
@@ -87,7 +87,7 @@ phase_function = none
 [FOV]
 
 camera_model = footprint
-
+visits_query = SELECT ra1,dec1,ra2,dec2,ra3,dec3,ra4,dec4,racenter,deccenter FROM observations WHERE observationId = ?
 ; # Choose between circular or actual camera footprint, including chip gaps.
 ; # Options: circle, footprint.
 ; camera_model = circle
