### 2025-04-01 9ae41e7
Added config file linkingfilter.ssp_number_observations into PPMiniDifi
Added linkingfilter.ssp_number_observations into MiniDifi function hasTracklet. Now the function is no longer hard coded for 2 observations to make a tracklet and depends on config variable ssp_number_observations. Updated test_PPLinkingFilter unit test and ran notebook demo_miniDifiValidation to check these changes with the brute force method to ensure the linking filter still works the same way.
