+++ tests/sorcha/test_PPLinkingFilter.py
@@ -111,6 +111,7 @@ def test_PPLinkingFilter_discoveryChances():
         objectId="ssObjectId",
         maxdt_minutes=max_time_separation * 24 * 60,
         minlen_arcsec=min_angular_separation,
+        min_observations = min_observations,
         window=min_tracklet_window,
         nlink=min_tracklets,
         p=detection_efficiency,
@@ -158,6 +159,7 @@ def test_PPLinkingFilter_nlink1():
         objectId="ssObjectId",
         maxdt_minutes=max_time_separation * 24 * 60,
         minlen_arcsec=min_angular_separation,
+        min_observations = min_observations,
         window=min_tracklet_window,
         nlink=min_tracklets,
         p=detection_efficiency,
