### 2025-04-03 073bafc
Changed linkingfilter config class checks
Changed linkingfilter config class checks to so that ssp_separation_threshold can be zero when ssp_number_observations =1
