### 2025-04-07 6813bc2
Edited DES config file and added DES discovery filter
Edited DES config file comments and added a very rough draft of a DESDiscoveryFilter
