+++ src/sorcha/data/DES/DES_config_file.ini
@@ -4,8 +4,8 @@
 # simulating a circular detector and approximating the chip gaps, removing all objects 
 # outside of a 1.1deg radius and 30% of objects within. 
 
-# The current DES sim has a circular fov and, uncertainties, randomisation, vignetting, 
-# trailing losses and fading function are turned off.
+# The current DES sim has uncertainties, randomisation, vignetting and 
+# trailing losses turned off.
 
 
 [INPUT]
@@ -72,7 +72,6 @@ observing_filters = r,g,i,z
 # Single float: applies same saturation limit to observations in all filters.
 # Comma-separated list of floats: applies saturation limit per filter, in order as
 # given in observing_filters keyword.
-#approximately 18
 bright_limit = 18.0
 
 
@@ -82,14 +81,17 @@ bright_limit = 18.0
 # file must contain the columns needed to calculate the phase function.
 # Options: HG, HG1G2, HG12, linear, none.
 phase_function = none
-#turn off for now 
 
 [FOV]
 
+# Choose between circular or actual camera footprint, including chip gaps. 
+# Options: circle, footprint.
+# for DES footprint a visits database is needed on the command line and a visits_query in this config section
+# the visits database is a sqlite database of the center and four corners of ccds in each exposure of the survey.
 camera_model = footprint
 visits_query = SELECT ra1,dec1,ra2,dec2,ra3,dec3,ra4,dec4,racenter,deccenter FROM observations WHERE observationId = ?
-; # Choose between circular or actual camera footprint, including chip gaps.
-; # Options: circle, footprint.
+
+# below is for a circular approximation of DES footprint
 ; camera_model = circle
 
 ; # Fraction of detector surface area which contains CCD -- simulates chip gaps
@@ -104,14 +106,20 @@ visits_query = SELECT ra1,dec1,ra2,dec2,ra3,dec3,ra4,dec4,racenter,deccenter FRO
 [FADINGFUNCTION]
 
 # Uses probability calculation from Bernardinelli et al (2022)
-# calculation is done per observation and so only need to turn ff on with no parameters needed
+# For DES, calculation is done per observation and so only need to turn fading function on with no parameters needed.
+# pointing sqlite database must include constants c and k. (can be seen in pointing_sql_query in the INPUT section.)
 fading_function_on = True
 
 
 [LINKINGFILTER]
+# For DES a single detection was considered a tracklet detection and 
+# a minimum of 7 unique nights of detection are needed to classify an object over the entire survey time.
+# The default Sorcha LinkingFilter is used to filter these detections and 
+# then a Discovery Filter is used for more specific requirements for discovering objects with DES. see Bernardinelli et al (2022)
 
-# SSP detection efficiency. Which fraction of the objects will
-# the automated Rubin Solar System Processing (SSP) pipeline successfully link? Float.
+
+# # SSP detection efficiency. Which fraction of the objects will
+# # the automated Rubin Solar System Processing (SSP) pipeline successfully link? Float.
 SSP_detection_efficiency = 1
 
 
@@ -121,11 +129,11 @@ SSP_number_observations = 1
 
 # Minimum separation (in arcsec) between two observations of an object required 
 # for the linking software to distinguish them as separate and therefore as a valid tracklet.
-SSP_separation_threshold = 0.0002
+SSP_separation_threshold = 0
 
 # Maximum time separation (in days) between subsequent observations in a tracklet. 
 # Default is 0.0625 days (90mins).
-SSP_maximum_time = 0.5
+SSP_maximum_time = 0
 
 # Number of tracklets for detection. How many tracklets are required to classify
 # an object as detected? 
[…]
