+++ docs/postprocessing.rst
@@ -206,11 +206,10 @@ LSSTCometActivity Class
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 Inside the `Sorcha add-ons  GitHub repository <https://github.com/dirac-institute/sorcha-addons>`_, we provide an example of a comet activity class. To use this function, in the :ref:`CPP` file, the user must provide a dust falling exponential value (k), the :math:`{af\rho}`, the product of albedo, the filling factor of grains within the observer field of view, and the linear radius of the field of view (typically selected to be 10,000 km) for the comet when it is located at a heliocentric distance of  1 au (:math:`{af\rho_1}`), Observing filter of the observation (optFilter), Apparent magnitude in the input filter of the comet nucleus adding up all of the counts in the trail (TrailedSourceMag). If these are not provided then the software will produce an error message. We have an `example Jupyter notebook <notebooks/demo_Cometary_Activity.ipynb>`_ demonstrating the LSSTCometActivity class built into `Sorcha add-ons package  <https://github.com/dirac-institute/sorcha-addons>`_. To use this prescription the **comet_activity** :ref:`configuration file<configs>` variable should be set to **lsst_comet**. The :math:`{af\rho(r)}` at the given observation is then estimated using the equation:
-.. math::{af\rho(r)} = {af\rho_1} r^kf({\alpha})
 
+.. math::
+  af\rho(r) = af\rho_1 \cdot r^k \cdot f(\alpha)
 Where :math:`{f(\alpha)}` is the Halley-Marcus phase function at phase angle (α) from `Schleicher <https://asteroid.lowell.edu/comet/dustphase/>`_.
-
-
 
 
 Rotational Lightcurve Effects
