### 2025-04-08 0a29900
Fixed if statement in miniDifi to not count duplicate detections
Fixed if statement in miniDifi to not count duplicate detections using set function. Added another unit test to check case where not enough observations are linked to count as a tracklet.
