### 2025-04-09 9a4ef06
Added DESCuts and DESDiscoveryFilter to DES sim
added DESCuts. This will remove objects outside of a chosen distance range or motion range if value is set in LINKINGFILTER section of the config file.
added another linkingfilter run in DES.py to remove objects that don't fit a approximate triplet criteria from Bernardinelli et al (2022)
added DESDiscoveryFilter. this checks that objects detected fit the criteria of a detected object from Bernardinelli et al (2022).

Note: all these new features still need unit tests added
