### 2025-04-09 11c2365
Added unit tests for DESCuts, DESDiscoveryFilter and sorchaConfigs
