### 2025-04-09 ea7b082
Added unit test for DESDetectionProbability
