+++ src/sorcha/modules/DESApplyFOVFilter.py
@@ -3,7 +3,7 @@ import numpy as np
 from astropy.coordinates import SkyCoord
 
 from sorcha.utilities.sorchaModuleRNG import PerModuleRNG
-
+from sorcha.modules.PPApplyFOVFilter import PPCircleFootprint,PPGetSeparation,PPSimpleSensorArea
 
 def DESApplyFOVFilter(observations, sconfigs, module_rngs, footprint=None, verbose=False):
     """
@@ -62,112 +62,3 @@ def DESApplyFOVFilter(observations, sconfigs, module_rngs, footprint=None, verbo
 
     return observations
 
-
-def PPGetSeparation(obj_RA, obj_Dec, cen_RA, cen_Dec):
-    """
-    Function to calculate the distance of an object from the field centre.
-
-    Parameters
-    -----------
-    obj_RA : float
-        RA of object in decimal degrees.
-
-    obj_Dec: float
-        Dec of object in decimal degrees.
-
-    cen_RA : float
-        RA of field centre in decimal degrees.
-
-    cen_Dec : float
-        Dec of field centre in decimal degrees.
-
-    Returns
-    -----------
-    sep_degree : float
-        The separation of the object from the centre of the field, in decimal
-        degrees.
-
-    """
-
-    obj_coord = SkyCoord(ra=obj_RA, dec=obj_Dec, unit="deg")
-    cen_coord = SkyCoord(ra=cen_RA, dec=cen_Dec, unit="deg")
-
-    sep = obj_coord.separation(cen_coord)
-
-    return sep.degree
-
-
-def PPCircleFootprint(observations, circle_radius):
-    """
-    Simple function which removes objects which lay outside of a circle
-    of given radius centred on the field centre.
-
-    Parameters
-    -----------
-    observations : Pandas dataframe
-        dataframe of observations.
-
-    circle_radius : float
-        radius of circle footprint in degrees.
-
-    Returns
-    ----------
-    new_observations : Pandas dataframe
-        dataframe of observations with all lying beyond the circle radius dropped.
-
-    """
-
-    data_coords = SkyCoord(ra=observations["RA_deg"].values, dec=observations["Dec_deg"].values, unit="deg")
-
-    field_coords = SkyCoord(
-        ra=observations["fieldRA_deg"].values, dec=observations["fieldDec_deg"].values, unit="deg"
-    )
-
-    separations = data_coords.separation(field_coords).degree
-    observations["object_separation"] = separations
-
-    new_observations = observations[observations["object_separation"] < circle_radius]
-
-    new_observations.reset_index(drop=True, inplace=True)
-    new_observations = new_observations.drop("object_separation", axis=1)
-
-    return new_observations
-
-
-def PPSimpleSensorArea(ephemsdf, module_rngs, fillfactor=0.9):
-    """
-    Randomly removes a number of observations proportional to the
-    fraction of the field not covered by the detector.
-
-    Parameters
-    -----------
-    ephemsdf : Pandas dataframe
-        Dataframe containing observations.
-
-    module_rngs : PerModuleRNG
-        A collection of random number generators (per module).
-
-    fillfactor : float
-        fraction of FOV covered by the sensor.
-        Default = 0.9
-
-    Returns
-    ----------
-    ephemsOut : Pandas dataframe
-        Dataframe of observations with 1- fillfactor fraction of objects
-        removed per on-sky observation pointing.
-
-    """
-    # Set the module specific seed as an offset from the base seed.
-    rng = module_rngs.getModuleRNG(__name__)
-
-    n = len(ephemsdf)
-
-    randomNum = rng.random(n)
[…]
