### 2025-04-28 #7746 @mschwamb
@warrickball The AAS journals paper was just accepted
