### 2025-04-29 #7746 @augustfly
@warrickball From the AAS perspective, the primary requirement is the start of a JOSS review. Once that review is started, we will know the future JOSS article DOI. We can easily proceed with this information (DOI plus the JOSS metadata) with the preparation of the AAS Journal article. Let me know if you need more information. - gus
