### 2025-04-29 #7746 @augustfly
We really only need the author list, title, journal, year and DOI. As year => volume, we can make an assumption given the time of year and the strength (in my opinion) of the developer's interest in completing a review that (year = 2025; Volume 10). That leaves the DOI, which is set by the start of a formal JOSS review.

This leaves the issue (incremental month) and honestly, it is not necessary for us to include issue for the purposes of citation and linking. The DOI does it all for us.
