### 2025-04-30 d888830
Updated miniDifi changes discussed with Mario
