+++ src/sorcha/modules/PPMiniDifi.py
@@ -83,23 +83,17 @@ def hasTracklet(mjd, ra, dec, maxdt_minutes, minlen_arcsec, min_observations):
     nobs = len(ra)
     if nobs < min_observations:
         return False
-
     if min_observations == 1:  # for edge case where one observation is needed for a tracklet
         return True
     maxdt = maxdt_minutes / (60 * 24)
     minlen = minlen_arcsec / 3600
-    detection_indexes = []  # index of objects linked to eachother in pairs
-
     for i in range(nobs):
         for j in range(nobs):
             diff = mjd[i] - mjd[j]
             if diff > 0 and diff < maxdt:
                 sep = haversine_np(ra[i], dec[i], ra[j], dec[j])
                 if sep > minlen:
-                    detection_indexes.append(i)
-                    detection_indexes.append(j)  # store indexes of observations that are linked.
-
-                    if len(set(detection_indexes)) >= min_observations:
+                    if (i + 1) - (j) >= min_observations:
                         return True
 
     return False
