### 2025-05-01 #7746 @warrickball
For transparency, I'll note here potential conflicts that I've chosen to waive.

* @jessie-dotson shares co-authorship with some Sorcha authors via several planetary defense papers with many authors (e.g. [1](https://ui.adsabs.harvard.edu/abs/2022Icar..37414790R), [2](https://ui.adsabs.harvard.edu/abs/2022PSJ.....3..123R), [3](https://ui.adsabs.harvard.edu/abs/2024PSJ.....5..141R)).
* @Onoddil shares co-authorship with some Sorcha authors via [the Astropy 5.0 paper](https://ui.adsabs.harvard.edu/abs/2022ApJ...935..167A) and is also a member of the [LSST:UK collaboration](https://www.lsst.ac.uk/).

All represent large collaborations with limited overlap and we've agreed that none prevent the reviewers from impartially assessing Sorcha.
