### 2025-05-01 #8145 @editorialbot
[REVIEW]: Sorcha: A Solar System Survey Simulator for the Legacy Survey of Space and Time
**Submitting author:** @mschwamb (<a href="http://orcid.org/0000-0003-4365-1455">Megan E. Schwamb</a>)
**Repository:** https://github.com/dirac-institute/sorcha/
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.1
