### 2025-05-01 #8145 @warrickball
Hi @jessie-dotson & @Onoddil, and thanks again for agreeing to review (and so incredibly quickly!). This is the review thread for the paper. All of our correspondence will now mainly happen here.

Please read the "Reviewer instructions & questions" in the [first comment](https://github.com/openjournals/joss-reviews/issues/8145#issue-3034001320) above, and generate your checklists by commenting `@editorialbot generate my checklist` on this issue. As you go over the submission, please check off any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

The JOSS review is different from most other journals. We aim to work with the authors to help them meet our criteria instead of merely passing judgement on the submission. We also encourage reviewers to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#8145` so that the issue/PR is linked to this thread. Please also feel free to comment and ask questions on this thread. JOSS editors have found it better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

If it suits your workflow, you're welcome to assign yourself to this issue in the GitHub UI.

Finally, don't hesitate to ask any questions you might have about the process.
