### 2025-05-06 #8145 @augustfly
The AAS Astronomical Journal article DOI will be: 10.3847/1538-3881/add3ec
