### 2025-05-06 #8145 @mschwamb
DOI added  to the JOSS article
