### 2025-05-16 8bb41d5
First stab at unit test -- still incomplete
