### 2025-05-16 b347eb7
Update test_simulation_parsing.py
