### 2025-05-17 df78e7b
update docs to emphasize further massive asteroid perturbers are bad to simulate
update docs to emphasize further massive asteroid perturbers are bad to simulate

addresses #1148
