+++ docs/whatsorchadoesnotdo.rst
@@ -8,6 +8,7 @@ in the future. If you want to add any of these features into ``Sorcha``, please 
 
 Here is a short summary of the key effects not accounted for in ``Sorcha``:
 
+- Properly simulating the locations of the 16 massive asteroid perturbers in the main belt. Further details can be found :ref:`here<MAP>`.  
 - Changing phase curves due to changing viewing angles (impacts some inner Solar System objects)
 - Stellar crowding as a function of galactic latitude
 - Non-gravitational forces including cometary outgassing or Yarkovsky or YORP (Yarkovsky–O'Keefe–Radzievskii–Paddack) effect. Although not directly handled, you can input your own ephemeris files that account for these effects if required.  
@@ -21,3 +22,8 @@ Here is a short summary of the key effects not accounted for in ``Sorcha``:
    magnitude of the simulated objects  due to cometary activity, rotational light curves, cometary 
    outbursts, etc. We have some basic functionality already built for simple sinusoidal rotational 
    light curves and cometary activity. Further details can be found  :ref:`here<addons>`.
+
+.. warning::
+  If you simulate the orbits of 16 massive asteroid perturbers listed  :ref:`here<MAP>`, you will get **POOR results** with the internal ``Sorcha`` ephemeris generator because of how the n-body integration is set up. We recommend getting the positions of these asteroids from some other source and inputting them as an external ephemeris file.
+
+
