### 2025-05-19 61e25cc
add a mention about moving observatories
add a mention about moving observatories
