### 2025-05-20 #8145 @warrickball
Hi again @jessie-dotson & @Onoddil, I just wanted to check in to ask whether you'd had a chance to start your reviews? As a reminder, at JOSS we encourage you generate your checklists and comment as you work through them, even if you're just checking off one or two items when you have time. That gives the authors a chance to start responding while you review other things, and usually helps wrap up the review faster.
