### 2025-05-21 dirac-institute/sorcha#1150 @Onoddil
JOSS Review Ticket
An issue in which to track the JOSS review (openjournals/joss-reviews#8145) issues, since there aren't many, they're not particularly large, and it doesn't really need separate tickets.

- [x] 1. `.git/objects/pack` is quite large when you `git clone` the package, causing 140MB of the 240MB total clone download. Is there any way to reduce that size without affecting anything? Less of a problem since your pip/conda compiled packages are small, though.
- [x] 2. Following the installation documentation (in particular the `conda install` method), I wasn't able to then continue blindly through to running the demo notebooks, since there was no mention of `jupyter` as a dependency during the installation process. Would be nice to add a reference, either at the top of the notebooks documentation section or in the appropriate place of installation, of the dependency. I was able to confirm the notebooks after a `conda install -c conda-forge jupyter`, so it shouldn't be too complicated to add reference to, hopefully. I also note that `requirements.txt` contains `jupyter` so it should have been grabbed if I were installing a compiled version in the "usual" way.
- [x] 3. Similarly, a few of the notebooks have `sorcha-addons` as a dependency, but the installation process says that unless you're editing the documentation you can skip step 6 (only then installing pandocs/sorcha-addons). Recommend just clearing that up -- probably, if the addon package is small enough, just make it a dependency of `sorcha` and hide that from the user, especially since `requirements.txt` contains `sorcha-addons`. I assume it was just missed from the `conda` installation step in the documentation.
- [x] 4. I wasn't able to find reference in the documentation of how, when contributing to the software, I might run the automated test suite (under `tests/`). I assume that the `sorcha demo` commands are separate (finding no reference to the test suite in the output logs from those commands), and a larger end-to-end proof-of-concept test -- which I can still run even if I install from pip or through conda -- and I should be using e.g. `pytest` to run the CI tests. However, there is (that I can see) no current method by which I would ensure that my contribution passes the automated test suite prior to submitting a PR (and, in this particular instance, verify the checklist item for JOSS).
