+++ docs/outputs.rst
@@ -5,7 +5,7 @@ Outputs
 
 ``Sorcha`` outputs:
   * :ref:`Detections File <detections>` (list of all the detections of the input popuation made by the simulated survey
-  * (Optioaal) :ref:`Statistics (Tally) File <statsf>`  that provides a summary overview for the objects from the input population that were ''found'' in the simulated survey
+  * (Optional) :ref:`Statistics (Tally) File <statsf>`  that provides a summary overview for the objects from the input population that were ''found'' in the simulated survey
   * (Optional) :ref:`Ephemeris Output <ephem_output>` that provides the output from the :ref:`Ephemeris Generation<ephemeris_gen>`  
 
 .. image:: images/survey_simulator_flow_chart.png
@@ -95,10 +95,6 @@ Detections File: Basic Output Column Names, Formats, and Descriptions
 +------------------------------------+--------------+----------------------------------------------------------------------------------+
 | Obj_Sun_LTC_km                     | Float        | Object-sun light-time-corrected distance (km)                                    |
 +------------------------------------+--------------+----------------------------------------------------------------------------------+
-| object_linked                      | Boolean      | True/False whether the object passed the linking filter. See note below          |
-+------------------------------------+--------------+----------------------------------------------------------------------------------+
-| date_linked_MJD                    | Boolean      | MJD (TAI) Date the object was linked (if it was linked) See note below           |
-+------------------------------------+--------------+----------------------------------------------------------------------------------+
 
 .. note::
    All positions and velocities are in respect to J2000.
@@ -117,16 +113,18 @@ Example Detections File in Basic Format
 
 .. code-block::
 
-   ObjID,fieldMJD_TAI,fieldRA_deg,fieldDec_deg,RA_deg,Dec_deg,astrometricSigma_deg,optFilter,PSFMag,trailedSourceMag,PSFMagSigma,trailedSourceMagSigma,fiveSigmaDepth_mag,fiveSigmaDepthAtSource
-   S1000000a,61769.320619,163.87542090842982,-18.84327137012991,164.03771300000017,-17.58257500000004,2.9880927198448093e-06,r,19.667095021023798,19.655534004675797,0.006775654132479691,0.006755926588113991,23.86356436464961,23.839403736057715
-   S1000000a,61769.332335,163.87542090842982,-18.84327137012991,164.03840499999956,-17.583782000000177,3.0580983448792015e-06,i,19.654439857054346,19.651499866857677,0.008648382870172588,0.00861644095296432,23.50948086026021,23.485408367730255
-   S1000000a,61773.283672,163.33185289781585,-17.478349047859123,164.25272700000096,-17.970833000000166,2.8628267283501646e-06,g,19.605094385361397,19.59913996244041,0.004573058990569846,0.004562676340629368,24.412081324532746,24.40274105573913
-   S1000000a,61773.304607,163.33185289781585,-17.478349047859123,164.2535509999998,-17.972800999999485,2.8619239276501636e-06,r,19.60417845127433,19.610463241887746,0.005414938113316873,0.005396964439230442,24.142184414583568,24.132798535794453
-   S1000000a,61780.286672,163.70205228035468,-18.10471138055092,164.4364500000006,-18.561287999999216,3.106487369364405e-06,i,19.50224387218658,19.49961057650898,0.00996299590797273,0.009945212307287087,23.1343489868631,23.13059981155987
[…]
