### 2025-06-02 #8145 @jessie-dotson
> I've left a few codebase-related queries in a ticket over in `sorcha`. Just a few very minor issues that hopefully won't be very much to smooth over!
>
> On the paper itself, I've just got two further minor review comments:
>
> * [ ]  the review checklist has a "state of the field" section, requiring comparison with "other commonly-used packages", but this is currently missing. I think you can make a case for the paper arguing there _isn't_ a comparison that can be made, and that kinda being the point of the creation of this software package, so I'd be happy to waive this if I am allowed to, but I note that currently it doesn't meet the letter of, rather than the spirit of, the checklist.[ ]  another very minor point, there are a few oddly formatted references at the moment. It looks like there're three submitted citations (al. MJH, al. M, and JAK et al. respectively) that aren't parsing right, but I guess they're playing the age-old game of "how do I resolve the DOIs of simultaneous reviews". I mention them just in case they can and/or should be updated! There are also a couple of citations that don't use "X et al. YEAR", instead formatting as "X et al. (YEAR)", so while updating the submitted papers just catch those (YEAR)-formatted citations.

I agree with these two issues identified by the other reviewer about the manuscript.
