### 2025-06-03 f7136cb
Update outputs.rst
fix typo
