### 2025-06-03 663b980
Update outputs.rst
