### 2025-06-03 4be9f6c
Update outputs.rst
