### 2025-06-03 69e39da
Update outputs.rst
make updates/minor corrections
