### 2025-06-03 #8145 @Onoddil
Sorry, yes, that wasn't to say there were _no_ comparisons, just that they weren't comparisons to other "equivalent" packages (since those aren't equivalent, which is why `sorcha` exists). Basically it was more a slightly pedantic editorial question mark, but I think (and assume Jessie would agree as above) that a quick "and none of these are fit for LSST because XYZ so we made this package" would make the point more emphatically.

I was just, as I said, obeying the letter of the review by not feeling like I was allowed to tick the section box without approval, which is presumably one for @warrickball ultimately!
