### 2025-06-03 #8145 @warrickball
Hi! I'd say don't worry about the very oddly formatted references. I can figure that out in the final editorial touches. We do like papers to finish the review with the references all sorted out but there are citation situations where it's unreasonable to expect the authors and reviewers to wrangle the pipeline into shape.

What both @Onoddil and @jessie-dotson have said about "how this software compares to other commonly-used packages" is correct. The spirit of the guideline is certainly that it's a valid comparison to argue that there are no comparable packages.
