+++ docs/index.rst
@@ -20,11 +20,11 @@ This documentation site contains an installation guide, an overview of how ``Sor
 works, tutorials, and demonstration notebooks that show how each of the various components within ``Sorcha`` work and can be customized.
 
 .. seealso::
-   For a more detailed description of ``Sorcha`` and how it works, please see  `Merritt et al. (in press) <https://www.dropbox.com/scl/fi/ku3s6aqpo6p8cambfclqi/Merritt_et_al_sorcha_paper.pdf?rlkey=07daa4q51nvr1tf7b818s3fy4&dl=0>`_  and `Holman et al. (in press) <https://www.dropbox.com/scl/fi/uofk7nz64a9xzf0aszmtx/sorcha_ephemeris_generation_paper.pdf?rlkey=2xmzwxar7qj2tj48qpqq3xcae&dl=0>`_.
+   For a more detailed description of ``Sorcha`` and how it works, please see  `Merritt et al. (in press) <https://arxiv.org/abs/2506.02804>`_  and `Holman et al. (in press) <https://arxiv.org/abs/2506.02140>`_.
 
 .. warning::
-   The code is still under going JOSS (Journal of Open Source Software) review, but we expect minimal changes as that finalizes. The code in the repository has been validated (see the :ref:`various validation notebooks we provide <demonotebooks>`).
-   We will release ``Sorcha`` v1.0 on PyPI and conda-forge soon - do double check the version from conda-forge/pypi. If your version is less than v1.0 uninstall and install via the development instructions using the source code directly from the Github repo. 
+   ``Sorcha`` v1.0 has been released on conda-forge and pypi. The code is still in the JOSS (Journal of Open Source Software) review process, but we expect minimal changes, but there could be a minor bug fix or two as that finalizes. So do install the latest version before finalizing your publication ready simulation output. The code in the repository has been validated (see the :ref:`various validation notebooks we provide <demonotebooks>` for LSST-like moving object discovery/linking).
+    
 
 
 What is Sorcha?
