### 2025-06-04 f1ff187
Update cite.rst
