### 2025-06-04 b83c9b4
updated DESCuts and DESDiscoveryFilter
made DESCuts more efficent  by taking the square of distance instead of square rooting.

Updated DESDiscoveryFilter to use the arccut and triplet functions from DESTNOSIM
