+++ src/sorcha/modules/DESDiscoveryFilter.py
@@ -1,35 +1,68 @@
 import numpy as np
 import astropy.units as u
+import numba
 
+from numba.typed import List
 
 ## filter for the DES object discovery requirements.
 #    An ARCCUT limit (has to be at least 2 objects not in a triplet discovery season)
 #    the observation of the discovery triplets have to be within 60/90 days (depending on distance) of eachother
+#    compute_arccut and compute_triplet are from DESTNOSIM: https://github.com/bernardinelli/DESTNOSIM
+
+bound = ((50 * u.au).to(u.km).value) ** 2  # square of the boundary 50au
 
 
-def DESDiscoveryFilter(observations):
+def DESDiscoveryFilter(
+    observations,
+    objectId="ObjID",
+    mjdTime="midPointTai",
+    x_km="Obj_Sun_x_LTC_km",
+    y_km="Obj_Sun_y_LTC_km",
+    z_km="Obj_Sun_z_LTC_km",
+):
+    """
+    Filter for the DES object discovery requirements. This filter checks for an ARCCUT limit (has to be at least 2 objects not in a triplet discovery season) and 
+    that the observation of the discovery triplets have to be within 60/90 days (depending on distance) of eachother.
+    Parameters
+    -----------
+
+    observations : Pandas dataframe
+        Dataframe of observations containing midPointTai and distance from Sun in cartesian co-ordinates.
+
+    Returns
+    ----------
+    observations : Pandas dataframe)
+        Modified 'observations' dataframe without observations that could not be observed.
+
+    """
+
     discovered_indices = []
+    i = np.argsort(observations[objectId]) # index of objects sorted  
+    _ , idx = np.unique(observations[objectId][i], return_index=True) # making an idx for each unique object
 
-    for object in observations["ObjID"].unique():
-        obj = observations[observations["ObjID"] == object]
-        bound = (50 * u.au).to(u.km).value
-        distance = np.sqrt(
-            observations["Obj_Sun_x_LTC_km"].values ** 2
-            + observations["Obj_Sun_y_LTC_km"].values ** 2
-            + observations["Obj_Sun_z_LTC_km"].values ** 2
-        )
-        if all(distance >= bound):
+    splits = np.split(i, idx[1:]) # splitting the objects into their detections 
+
+    for _ , obsv_indices in enumerate(splits): # loop for each object
+        thisObsv = observations[[objectId, mjdTime, x_km, y_km, z_km]][obsv_indices]
+        thisObsv.dtype.names = [objectId, mjdTime, x_km, y_km, z_km]
+
+        distance_sq = thisObsv[x_km] ** 2 + thisObsv[y_km] ** 2 + thisObsv[z_km] ** 2 # should i just calculate for one detection 
+        if any(distance_sq >= bound):  # boundary condition for triplet detection (depends on distance)
             window = 90
         else:
             window = 60
+        # sort the objects observations by time mjd
+        i_times = np.argsort(thisObsv[mjdTime]) 
+        thisObsv = thisObsv[i_times] 
 
-        if not arccut(np.array(obj["fieldMJD_TAI"])):
-            continue
+        # check cinditions are meet for detection. 
+        if not compute_arccut(thisObsv[mjdTime]) > 0.5 * 365.25 or not compute_triplet(
+            thisObsv[mjdTime], window
+        ):
+            continue # if conditions are not met skip object
 
-        if not triplet(np.array(obj["fieldMJD_TAI"]), window):
-            continue
-
-        discovered_indices.extend(obj.index.tolist())
+        discovered_indices.extend(obsv_indices)
[…]
