### 2025-06-04 2049d51
updated unit test and fixed bugs for DESDiscoveryFilter
