### 2025-06-04 69b943b
Update DESDiscoveryFilter.py
