### 2025-06-04 2c4de8d
Update DES.py
