### 2025-06-05 3b9ecc2
added unit test and some changes to DESDiscoveryFilter
added unit test for multiple objects passed through

cleaned up and made Discovery filter more efficient
