### 2025-06-27 a18ff75
updated unit test for Discovery filter
