### 2025-06-27 d9bdad9
Changes to motionCut and sorchaConfigs
made motioncut more efficient and fixed name of variables and error message in filtersConfigs
