### 2025-06-27 38b4e65
Added check for DESfadingfunction  in PPReadPointingDatabase so code will fail gracefully
