### 2025-06-30 05dc6a5
Fixing edge cases and footprint
large amount of cleaning up code in the DES branch.

fov
changed how fov works in Sorcha. now DES can use fov in the two same ways as LSST. the visits_footprint is the new fov.camera_model for per observation ccd locations. currently only DES can do this. This involved cleaning up a large portion of the code and making the functions more modular. DESApplyFOVFilter is no longer needed and the PPApplyFOVFilter is now the wrapper for all the fov filters. DESFootprintFilter has now be replaced with PPVisitsFootprintFilter. Updated the checks for visits_footprint to 
[…]
