+++ tests/sorcha/test_sorchaConfigs.py
@@ -504,7 +504,7 @@ def test_fovConfigs_inlist():
         test_configs = fovConfigs(**fov_configs)
     assert (
         error_text.value.code
-        == "ERROR: value fake_model for config parameter camera_model not recognised. Expecting one of: ['circle', 'footprint', 'none']."
+        == "ERROR: value fake_model for config parameter camera_model not recognised. Expecting one of: ['circle', 'footprint', 'visits_footprint', 'none']."
     )
 
 
@@ -558,13 +558,14 @@ def test_fovConfigs_circle_mandatory():
         error_text.value.code
         == 'ERROR: either "fill_factor" or "circle_radius" must be specified for circular footprint.'
     )
-def test_fovConfigs_DES():
+def test_fovConfigs_visits_footprint():
     """
-    Makes sure the code fails when using DES and having an edge thresh
+    Makes sure the code fails when using visits_footprint and having an edge thresh
     """
 
     fov_configs = correct_fov.copy()
     fov_configs["survey_name"] = "DES"
+    fov_configs["camera_model"] = "visits_footprint"
 
     fov_configs["visits_query"] = "something"
 
@@ -572,7 +573,7 @@ def test_fovConfigs_DES():
         test_configs = fovConfigs(**fov_configs)
     assert (
         error_text.value.code
-        == "ERROR: footprint_edge_threshold supplied in config file But DES doesn't use edge threshold"
+        == "ERROR: footprint_edge_threshold supplied in config file But visits footprint does not use edge threshold"
     )
 
     fov_configs["visits_query"] = None
