+++ src/sorcha/utilities/sorchaArguments.py
@@ -43,6 +43,9 @@ class sorchaArguments:
     pplogger = None
     """The Python logger instance"""
 
+    visits = None
+    """visits database. a sqlite database containing the camera footprint and on-sky location for each ccd per observation"""
+
     def __init__(self, cmd_args_dict=None):
         self.pplogger = logging.getLogger(__name__)
         if cmd_args_dict is not None:

[…]
