### 2025-06-30 5136dfd
fixed bug
if obs id had no corresponding visitsId for ccd an error would occur. will now skip fieldId if no ccds supplied
