### 2025-07-04 017a56c
refactored applyVisitsFootprint function
refactored applyVisitsFootprint function to improve computation time and output detector id
