+++ src/sorcha/modules/PPVisitsFootprintFilter.py
@@ -1,97 +1,91 @@
 import logging
 
-from shapely.geometry import Point, Polygon, MultiPolygon
+from shapely.geometry import Point, Polygon
 import sqlite3
 from scipy.spatial import KDTree
-
+import numpy as np
 logger = logging.getLogger(__name__)
 
-
+@profile
 def applyVisitsFootprint(
     field_df, query, visits_filename, ra_name="RA_deg", dec_name="Dec_deg", fieldId="FieldID"
 ):
     """
     Determine whether detections fall on the sensors defined by the
-    footprint. Also returns the an ID for the sensor a detection is made
+    footprint. Also returns the ID for the sensor a detection is made
     on.
 
     Parameters
-    -----------
-    field_df : Pandas dataframe
-        Dataframe containing detection information with pointings.
-
-    ra_name : string, optional
-        "field_df" dataframe's column name for object's RA
-        for the given observation. Default = "RA_deg" [units: degrees]
-
-    dec_name : string, optional
-        "field_df" dataframe's column name for object's declination
-        for the given observation. Default = "Dec_deg" [units: dgrees]
-
-    ra_name_field : string, optional
-        "field_df" dataframe's column name for the observation field's RA
-        Default = "fieldRA_deg" [units: degrees]
-
-    dec_name_field : string, optional
-        "field_df" dataframe's column name for the observation field's declination
-        Default = "fieldDec_deg" [Units: degrees]
-
-
+    ----------
+    field_df : pandas.DataFrame
+        DataFrame containing detection information with pointings.
+    query : str
+        SQL query string for visits database. name columns as (llcra, llcdec, lrcra, lrcdec, urcra, urcdec, ulcra, ulcdec, ra_centre, dec_centre, detector)
+    visits_filename : str
+        Path to SQLite database containing footprint data.
+    ra_name, dec_name, fieldId : str
+        Column names in field_df for RA, Dec, and field ID respectively.
 
     Returns
-    ----------
-    detected : array
-        Indices of rows in field_df which fall on the sensor(s).
-
-    detectorID :array
-        name of the detector that the object falls on.
+    -------
+    detected_list : list
+        List of indices in field_df that fall on a sensor footprint.
+    detector_id_list : list
+        List of detector IDs corresponding to each detected object.
     """
+    detected_indices = set()
+    detector_for_index = {}
 
-    detected_index = set()
-
-    # sqlite database connection
+# open visits database file
     with sqlite3.connect(visits_filename) as conn:
+        conn.row_factory = sqlite3.Row
         cursor = conn.cursor()
-        # taking all non duplicate exposure ids (Some footprints have a chance of no objects in them and so polygons won't be made of them when using the ids from the detection pd)
-        field_df_no_duplicates = field_df.drop_duplicates(subset=[fieldId], keep="first")
-        # going through each polygon in a loop
 
-        for obs_id in field_df_no_duplicates[fieldId]:
+        grouped_fieldId = field_df.groupby(fieldId) #group objects by fieldId maybe using np.argsort instead would be faster like in miniDifi?
+        for obs_id, points_df in grouped_fieldId:
+            rows = cursor.execute(query, (obs_id,)).fetchall() # execute query on visits database
+            if not rows:
+                continue # if no ccds skip
 
-            # executing query for this observation id
-            rows = cursor.execute(query, (obs_id,)).fetchall()
-            if rows != []: # next step if there is a matching id in the visits database
-                # filtering points for given observation Id
-                check_points = field_df[field_df[fieldId] == obs_id]
-                points_query = [
-                    list([ra_val, dec_val])
-                    for ra_val, dec_val in zip(check_points[ra_name], check_points[dec_name])
-                ]
-                # creating kd tree of center of each ccd and finding 3 closest ccds to each point
[…]
