+++ src/sorcha/modules/PPVisitsFootprintFilter.py
@@ -6,7 +6,7 @@ from scipy.spatial import KDTree
 import numpy as np
 logger = logging.getLogger(__name__)
 
-@profile
+
 def applyVisitsFootprint(
     field_df, query, visits_filename, ra_name="RA_deg", dec_name="Dec_deg", fieldId="FieldID"
 ):
