### 2025-07-07 715f468
Added unit test for visitsFootprintFilter
Added unit test for visitsFootprintFilter and cleaned some small bugs and wording of comments
