### 2025-07-07 #8145 @mschwamb
@warrickball DOIs came in from everything in the past two weeks. Had to deal with exam marking and getting a press release out on the code and science results. So I only now have time. It's on my todo list for this week to get to @Onoddil's review. I don't see any ticket in the sorcha repo for @jessie-dotson's review.
