### 2025-07-15 545e98d
changes to DESCuts and changing code to be more in line with dp1
renamed DESCuts to PPDistanceandMotionCuts and made changes to parameter names.
changed Visits database to also output limiting mag per ccd chip.
