+++ src/sorcha/modules/PPApplyFOVFilter.py
@@ -59,11 +59,11 @@ def PPApplyFOVFilter(observations, sconfigs, module_rngs, visits=None, footprint
 
     if sconfigs.fov.camera_model == "visits_footprint":
         verboselog("Applying sensor footprint filter...")
-        onSensor, detectorIDs = PPVisitsFootprint(observations, sconfigs.fov.visits_query, visits)
+        onSensor, detectorIDs, lim_mag = PPVisitsFootprint(observations, sconfigs.fov.visits_query, visits)
 
         observations = observations.iloc[onSensor].copy()
         observations["detectorID"] = detectorIDs
-
+        observations["fieldFiveSigmaDepth_mag"] = lim_mag 
         observations = observations.sort_index()
 
     if sconfigs.fov.camera_model == "circle":
