+++ src/sorcha/modules/PPReadPointingDatabase.py
@@ -59,7 +59,7 @@ def PPReadPointingDatabase(bsdbname, observing_filters, dbquery, surveyname, fad
         dfo["observationMidpointMJD_TAI"] = dfo["observationStartMJD_TAI"] + (
             (dfo["visitTime"] / 2.0) / 86400.0
         )
-    elif surveyname in ["DES", "des"]:
+    elif surveyname in ["DES", "des","lsst","LSST"]:
         dfo["observationStartMJD_TAI"] = dfo["observationMidpointMJD_TAI"] - (
             (dfo["visitExposureTime"] / 2.0) / 86400.0
         )

[…]
