### 2025-07-17 b38dfde
Added a dp1 simulation option
sorcha init - added a dp1 config file option

added demo notebook to download sqlite databases for dp1 simulations

Updated fading function to take per chip limiting mags if available for dp1

updated code to include survey name dp1 with all the checks and balances needed for that.

updated and fixed unit tests for these changes.
