+++ src/sorcha/DP1.py
@@ -1,0 +1,385 @@
+#!/usr/bin/python
+
+import sys
+import time
+import numpy as np
+import argparse
+import os
+import logging
+
+from sorcha.ephemeris.simulation_driver import create_ephemeris
+from sorcha.ephemeris.simulation_setup import precompute_pointing_information
+
+from sorcha.modules.PPReadPointingDatabase import PPReadPointingDatabase
+from sorcha.modules.PPLinkingFilter import PPLinkingFilter
+from sorcha.modules.PPTrailingLoss import PPTrailingLoss
+from sorcha.modules.PPBrightLimit import PPBrightLimit
+from sorcha.modules.PPCalculateApparentMagnitude import PPCalculateApparentMagnitude
+from sorcha.modules.PPApplyFOVFilter import PPApplyFOVFilter
+from sorcha.modules.PPSNRLimit import PPSNRLimit
+from sorcha.modules import PPAddUncertainties, PPRandomizeMeasurements
+from sorcha.modules import PPVignetting
+from sorcha.modules.PPFadingFunctionFilter import PPFadingFunctionFilter
+from sorcha.modules.PPFaintObjectCullingFilter import PPFaintObjectCullingFilter
+
+
+from sorcha.modules.PPMatchPointingToObservations import PPMatchPointingToObservations
+from sorcha.modules.PPMagnitudeLimit import PPMagnitudeLimit
+from sorcha.modules.PPOutput import PPWriteOutput, PPIndexSQLDatabase
+from sorcha.modules.PPGetMainFilterAndColourOffsets import PPGetMainFilterAndColourOffsets
+from sorcha.modules.PPFootprintFilter import Footprint
+from sorcha.modules.PPStats import stats
+
+from sorcha.readers.CombinedDataReader import CombinedDataReader
+from sorcha.readers.CSVReader import CSVDataReader
+from sorcha.readers.EphemerisReader import EphemerisDataReader
+from sorcha.readers.OrbitAuxReader import OrbitAuxReader
+
+from sorcha.activity.activity_registration import update_activity_subclasses
+from sorcha.lightcurves.lightcurve_registration import update_lc_subclasses
+
+from sorcha.utilities.sorchaArguments import sorchaArguments
+from sorcha.utilities.sorchaConfigs import sorchaConfigs, PrintConfigsToLog
+from sorcha.utilities.sorchaCommandLineParser import sorchaCommandLineParser
+from sorcha.utilities.fileAccessUtils import FindFileOrExit
+from sorcha.utilities.citation_text import cite_sorcha
+from sorcha.utilities.sorchaGetLogger import sorchaGetLogger
+
+
+def cite():  # pragma: no cover
+    """Providing the bibtex, AAS Journals software latex command, and acknowledgement
+    statements for Sorcha and the associated packages that power it.
+
+    Parameters
+    ----------
+    None
+
+    Returns
+    --------
+    None
+    """
+    cite_sorcha()
+
+
+def mem(df):
+    """
+    Memory utility function that returns back how much memory the inputted pandas dataframe is using
+    Parameters
+    ------------
+    df : pandas dataframe
+
+    Returns
+    -----------
+    usage : int
+
+    """
+
+    usage = df.memory_usage(deep=True).sum()
+    for k, v in df.attrs.items():
+        usage += v.nbytes
+    return usage
+
+
+def runDP1Simulation(args, sconfigs):
+    """
+    Runs the post processing survey simulator functions that apply a series of
+    filters to bias a model Solar System small body population to what the
+    Vera C. Rubin Observatory Legacy Survey of Space and Time would observe.
+
+    Parameters
+    -----------
+    args : dictionary or `sorchaArguments` object
+        dictionary of command-line arguments.
+
[…]
