+++ src/sorcha_cmdline/run.py
@@ -154,6 +154,7 @@ def execute(args):
         update_lc_subclasses,
     )
     from sorcha.DES import runDESSimulation
+    from sorcha.DP1 import runDP1Simulation
     import sys, os
 
     # Extract the output file path now in order to set up logging.
@@ -204,7 +205,7 @@ def execute(args):
             "ERROR: cmd line arg --vd, --visits-db and config fov varible visits_query must both be specified"
         )
 
-    if cmd_args["surveyname"] in ["rubin_sim", "RUBIN_SIM","lsst", "LSST"]:
+    if cmd_args["surveyname"] in ["rubin_sim", "RUBIN_SIM"]:
         try:
             args = sorchaArguments(cmd_args)
         except Exception as err:
@@ -216,6 +217,30 @@ def execute(args):
             pplogger.error(err)
             sys.exit(err)
         runLSSTSimulation(args, sconfigs)
+    elif cmd_args["surveyname"] in ["LSST", "lsst"]:
+        pplogger.error(
+            "ERROR: The LSST has not started yet Current allowed surveys are: {}".format(
+                ["rubin_sim", "RUBIN_SIM"]
+            )
+        )
+        sys.exit(
+            "ERROR: The LSST has not started. Current allowed surveys are: {}".format(
+                ["rubin_sim", "RUBIN_SIM"]
+            )
+        )
+    elif cmd_args["surveyname"] in ["DP1", "dp1"]:
+        try:
+            args = sorchaArguments(cmd_args)
+        except Exception as err:
+            pplogger.error(err)
+            sys.exit(err)
+        try:
+            args.validate_arguments()
+        except Exception as err:
+            pplogger.error(err)
+            sys.exit(err)
+
+        runDP1Simulation(args, sconfigs)
     elif cmd_args["surveyname"] in ["DES", "des"]:
         try:
             args = sorchaArguments(cmd_args)
