### 2025-07-17 0d62100
Update test_sorcha_copy_configs.py
