+++ tests/sorcha/test_sorcha_copy_configs.py
@@ -73,11 +73,13 @@ def test_parse_file_selection():
     test_rubin_circle = parse_file_selection("1")
     test_rubin_footprint = parse_file_selection("2")
     test_rubin_known = parse_file_selection("3")
-    test_all = parse_file_selection("4")
+    test_dp1_visits = parse_file_selection("4")
+    test_all = parse_file_selection("5")
 
     assert test_rubin_circle == "rubin_circle"
     assert test_rubin_footprint == "rubin_footprint"
     assert test_rubin_known == "rubin_known"
+    assert test_dp1_visits == "dp1_visits"
     assert test_all == "all"
 
     # test error messages
@@ -92,5 +94,5 @@ def test_parse_file_selection():
 
     assert (
         e2.value.code
-        == "Input could not be converted to a valid integer. Please input an integer between 1 and 4."
+        == "Input could not be converted to a valid integer. Please input an integer between 1 and 5."
     )
