+++ src/sorcha/modules/PPVisitsFootprintFilter.py
@@ -4,7 +4,9 @@ from shapely.geometry import Point, Polygon
 import sqlite3
 from scipy.spatial import KDTree
 import numpy as np
+
 logger = logging.getLogger(__name__)
+
 
 def PPVisitsFootprint(
     field_df, query, visits_filename, ra_name="RA_deg", dec_name="Dec_deg", fieldId="FieldID"
@@ -36,25 +38,29 @@ def PPVisitsFootprint(
     detector_for_index = {}
     lim_mag_list = {}
 
-# open visits database file
+    # open visits database file
     with sqlite3.connect(visits_filename) as conn:
         conn.row_factory = sqlite3.Row
         cursor = conn.cursor()
 
-        grouped_fieldId = field_df.groupby(fieldId) #group objects by fieldId maybe using np.argsort instead would be faster like in miniDifi?
+        grouped_fieldId = field_df.groupby(
+            fieldId
+        )  # group objects by fieldId maybe using np.argsort instead would be faster like in miniDifi?
         for obs_id, points_df in grouped_fieldId:
-            rows = cursor.execute(query, (obs_id,)).fetchall() # execute query on visits database
+            rows = cursor.execute(query, (obs_id,)).fetchall()  # execute query on visits database
             if not rows:
-                continue # if no ccds skip
+                continue  # if no ccds skip
 
-            points_query = np.array(list(zip(points_df[ra_name], points_df[dec_name]))) # number list of ra and of objects
+            points_query = np.array(
+                list(zip(points_df[ra_name], points_df[dec_name]))
+            )  # number list of ra and of objects
 
             ccd_centers = [(row["ra_centre"], row["dec_centre"]) for row in rows]
             ccd_tree = KDTree(ccd_centers)
-            if len(ccd_centers) > 1:                
-                k = min(3, len(ccd_centers)) # making sure theirs no error with 2 ccds
+            if len(ccd_centers) > 1:
+                k = min(3, len(ccd_centers))  # making sure theirs no error with 2 ccds
                 _, closest_ccd_indices = ccd_tree.query(points_query, k=k)
-                unique_ccd_indices = set(closest_ccd_indices.flatten()) #remove duplicates
+                unique_ccd_indices = set(closest_ccd_indices.flatten())  # remove duplicates
             else:
                 unique_ccd_indices = [0]
             # create polygons for closest ccds to points and record detector name/id
@@ -63,23 +69,24 @@ def PPVisitsFootprint(
             limmag = []
             for idx in unique_ccd_indices:
                 row = rows[idx]
-                corners = np.array([
-                    (row["llcra"], row["llcdec"]),
-                    (row["lrcra"], row["lrcdec"]),
-                    (row["urcra"], row["urcdec"]),
-                    (row["ulcra"], row["ulcdec"]),
-                    (row["llcra"], row["llcdec"])  # closing polygon
-                ])
+                corners = np.array(
+                    [
+                        (row["llcra"], row["llcdec"]),
+                        (row["lrcra"], row["lrcdec"]),
+                        (row["urcra"], row["urcdec"]),
+                        (row["ulcra"], row["ulcdec"]),
+                        (row["llcra"], row["llcdec"]),  # closing polygon
+                    ]
+                )
                 polygons.append(Polygon(corners))
                 detectors.append(row["detectorID"])
                 limmag.append(row["limMag_perChip"])
 
-
             # Create a list of Shapely Point objects for each possible detection
             points = [Point(point) for point in points_query]
 
-            for point_index, point in enumerate(points): # for every point
-                for poly_idx, polygon in enumerate(polygons): # for every ccd
+            for point_index, point in enumerate(points):  # for every point
+                for poly_idx, polygon in enumerate(polygons):  # for every ccd
[…]
