### 2025-07-17 a687526
added DES survey setup
