+++ src/sorcha/data/DES/DES_config_file.ini
@@ -1,185 +1,0 @@
-# Sorcha Configuration File
-# This configuration file will set up Sorcha to simulate what the DES would discover.
-# This version simulates objects lost to chip gaps between DECam's detectors by 
-# simulating a circular detector and approximating the chip gaps, removing all objects 
-# outside of a 1.1deg radius and 30% of objects within. 
-
-# The current DES sim has uncertainties, randomisation, vignetting and 
-# trailing losses turned off.
-
-
-[INPUT]
-
-# The simulation used for the ephemeris input. 
-# ar=ASSIST+REBOUND interal ephemeris generation 
-# external=providing an external input file from the command line
-# Options: "ar", "external"
-ephemerides_type = ar
-
-# Format for ephemeris simulation input file if a file is specified at the command line. 
-# This is also the format to which ephemeris files will be written out, if specified.
-# Options: csv, whitespace, hdf5
-eph_format = csv
-
-# Sorcha chunk size: how many objects should be processed at once?
-size_serial_chunk = 20000
-
-# Format for the orbit, physical parameters, and complex physical parameters input files.
-# Options: csv or whitespace
-aux_format = csv
-
-# SQL query for extracting data from the pointing database.
-pointing_sql_query = SELECT observationId, observationMidpointMJD as observationMidpointMJD_TAI , visitExposureTime, filter, fiveSigmaDepth as fieldFiveSigmaDepth_mag , fieldRA as fieldRA_deg, fieldDec as fieldDec_deg, c, k FROM observations order by observationId
-
-
-[SIMULATION]
-# Configuration for running the ASSIST+REBOUND ephemerides generator.
-
-# the field of view of our search field, in degrees
-ar_ang_fov = 1.3
-
-# the buffer zone around the field of view we want to include, in degrees
-ar_fov_buffer = 0.2
-
-# the "picket" is our imprecise discretization of time that allows us to move progress
-# our simulations forward without getting too granular when we don't have to.
-# the unit is number of days.
-ar_picket = 1
-
-# the obscode is the MPC observatory code for the provided telescope.
-ar_obs_code = W84
-
-# the order of healpix which we will use for the healpy portions of the code.
-# the nside is equivalent to 2**ar_healpix_order
-ar_healpix_order = 6
-
-
-[FILTERS]
-
-# Filters of the observations you are interested in, comma-separated.
-# Your physical parameters file must have H calculated in one of these filters
-# and colour offset columns defined relative to that filter.
-observing_filters = r,g,i,z
-
-
-[SATURATION]
-
-# Upper magnitude limit on sources that will overfill the detector pixels/have
-# counts above the non-linearity regime of the pixels where one can’t do 
-# photometry. Objects brighter than this limit (in magnitude) will be cut. 
-# Comment out for no saturation limit.
-# Two formats are accepted:
-# Single float: applies same saturation limit to observations in all filters.
-# Comma-separated list of floats: applies saturation limit per filter, in order as
-# given in observing_filters keyword.
-bright_limit = 18.0
-
-
-[PHASECURVES]
-
-# The phase function used to calculate apparent magnitude. The physical parameters input
-# file must contain the columns needed to calculate the phase function.
-# Options: HG, HG1G2, HG12, linear, none.
-phase_function = none
-
-[FOV]
-
-# Choose between circular or actual camera footprint, including chip gaps. 
-# Options: circle, footprint.
-# for DES footprint a visits database is needed on the command line and a visits_query in this config section
-# the visits database is a sqlite database of the center and four corners of ccds in each exposure of the survey.
[…]
