### 2025-08-08 e87e1ce
Added new demo for VisitsFootprintFilter
Added new demo for the VisitsFootprintFilter.

This demo uses newly created file DES_footprint.db containing two footprint ccd on sky location observations.

fixed small typo in DES config file
