+++ src/sorcha/data/survey_setups/DES_config_file.ini
@@ -86,7 +86,7 @@ camera_model = visits_footprint
 
 
 # the query needed for the visits footprint file. must include the ccd corners and centres, the detectors Id and the limiting mag. There will be a notebook that creates this in the future.
-visits_query = SELECT llcra, llcdec, lrcra, lrcdec, urcra, urcdec, ulcra, ulcdec, ra as ra_centre, dec as dec_centre, ccdVisitId as detectorID, magLim as limMag_perChip  FROM observations WHERE visitId = ?
+visits_query = SELECT llcra, llcdec, lrcra, lrcdec, urcra, urcdec, ulcra, ulcdec, ra as ra_centre, dec as dec_centre, detectorID, magLim as limMag_perChip  FROM observations WHERE visitId = ?
 
 
 
