### 2025-08-20 828a463
updates based on JOSS review
- started updated notebooks for bug fixes and requested reviewer changes
- documentation changes
- fixed mismatch of variable name with the doc strings
