### 2025-08-20 a3e7812
JOSS review updates
updates to notebooks to address minor comments from JOSS referees and make clearer the instructions for running the unit tests
