### 2025-08-24 d9263d8
JOSS and docstring udpates
updating the JOSS references to those with now published information.

Fixing docstring for #1161
