+++ JOSS_submission/paper.bib
@@ -253,17 +253,26 @@ archivePrefix = {arXiv},
 }
 
 
-@article{holman2025, 
-    year = {submitted},
-    month = {jan},
-    publisher = {The American Astronomical Society},
-    volume = {8},
-    number = {1},
-    pages = {},
-    author = {M. J. Holman  et al},
-    title = {Sorcha: Optimized Solar System Ephemeris Generation},
-    journal = {Astronomical Journal},
+@ARTICLE{holman2025,
+       author = {{Holman}, Matthew J. and {Bernardinelli}, Pedro H. and {Schwamb}, Megan E. and {Juri{\'c}}, Mario and {Oldag}, Drew and {West}, Maxine and {Napier}, Kevin J. and {Merritt}, Stephanie R. and {Fedorets}, Grigori and {Cornwall}, Samuel and {Kurlander}, Jacob A. and {Eggl}, Siegfried and {Kubica}, Jeremy and {Kiker}, Kathleen and {Murtagh}, Joseph and {Naidu}, Shantanu P. and {Chandler}, Colin Orion},
+        title = "{Sorcha: Optimized Solar System Ephemeris Generation}",
+      journal = {Astronomical Journal},
+     keywords = {Ephemerides, Small Solar System bodies, N-body simulations, Sky surveys, 464, 1469, 1083, 1464, Earth and Planetary Astrophysics, Instrumentation and Methods for Astrophysics, Computational Physics},
+         year = 2025,
+        month = aug,
+       volume = {170},
+       number = {2},
+          eid = {97},
+        pages = {97},
+          doi = {10.3847/1538-3881/ade435},
+archivePrefix = {arXiv},
+       eprint = {2506.02140},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025AJ....170...97H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
 
 @ARTICLE{hoover2022,
        author = {{Hoover}, Devin J. and {Seligman}, Darryl Z. and {Payne}, Matthew J.},
@@ -367,16 +376,23 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{kurlander2025,   
-    year = {submitted},
-    month = {jan},
-    publisher = {The American Astronomical Society},
-    volume = {},
-    number = {},
-    pages = {},
-    author = {J, A. Kurlander et al},
-    title = {Predictions of the LSST Solar System Yield: The Four Major Populations},
-    journal = {Astronomical Journal},
+@ARTICLE{kurlander2025,
+       author = {{Kurlander}, Jacob A. and {Bernardinelli}, Pedro H. and {Schwamb}, Megan E. and {Juri{\'c}}, Mario and {Murtagh}, Joseph and {Chandler}, Colin Orion and {Merritt}, Stephanie R. and {Nesvorn{\'y}}, David and {Vokrouhlick{\'y}}, David and {Jones}, R. Lynne and {Fedorets}, Grigori and {Cornwall}, Samuel and {Holman}, Matthew J. and {Eggl}, Siegfried and {Oldag}, Drew and {West}, Maxine and {Kubica}, Jeremy and {Yoachim}, Peter and {Moeyens}, Joachim and {Kiker}, Kathleen and {Buchanan}, Laura E.},
+        title = "{Predictions of the LSST Solar System Yield: Near-Earth Objects, Main Belt Asteroids, Jupiter Trojans, and Trans-Neptunian Objects}",
+      journal = {Astronomical Journal},
[…]
