### 2025-08-24 d596bd2
updates
update pyproject.toml to include most packages that are required for running the demo notebooks.

updated the instructions to refer to thi change that most required packages for the documentation are in installed ind dev mode
