### 2025-08-24 b23279f
Update notebooks.rst
update how to install the associated packages for the notebookes
