### 2025-09-01 0e8a6e3
update doc strings
