+++ src/sorcha/utilities/citation_text.py
@@ -18,10 +18,14 @@ def cite_sorcha(f=sys.stdout):
 
     f.write("\nSorcha: \n\n")
     f.write("@ARTICLE{2025AJ....170..100M,\n")
-    f.write("author = {{Merritt}, Stephanie R. and {Fedorets}, Grigori and {Schwamb}, Megan E. and {Cornwall}, Samuel and {Bernardinelli}, Pedro H. and {Juri{\'c}}, Mario and {Holman}, Matthew J. and {Kurlander}, Jacob A. and {Eggl}, Siegfried and {Oldag}, Drew and {West}, Maxine and {Kubica}, Jeremy and {Murtagh}, Joseph and {Jones}, R. Lynne and {Yoachim}, Peter and {Lyttle}, Ryan R. and {Kelley}, Michael S.~P. and {Moeyens}, Joachim and {Kiker}, Kathleen and {Naidu}, Shantanu P. and {Snodgrass}, Colin and {Matthews}, Shannon M. and {Chandler}, Colin Orion},\n")
+    f.write(
+        "author = {{Merritt}, Stephanie R. and {Fedorets}, Grigori and {Schwamb}, Megan E. and {Cornwall}, Samuel and {Bernardinelli}, Pedro H. and {Juri{'c}}, Mario and {Holman}, Matthew J. and {Kurlander}, Jacob A. and {Eggl}, Siegfried and {Oldag}, Drew and {West}, Maxine and {Kubica}, Jeremy and {Murtagh}, Joseph and {Jones}, R. Lynne and {Yoachim}, Peter and {Lyttle}, Ryan R. and {Kelley}, Michael S.~P. and {Moeyens}, Joachim and {Kiker}, Kathleen and {Naidu}, Shantanu P. and {Snodgrass}, Colin and {Matthews}, Shannon M. and {Chandler}, Colin Orion},\n"
+    )
     f.write('title = "{Sorcha: A Solar System Survey Simulator for the Legacy Survey of Space and Time}"\n,')
     f.write("journal = {\\aj},\n")
-    f.write("keywords = {Solar system astronomy, Small Solar System bodies, Astronomy software, Astronomical simulations, 1529, 1469, 1855, 1857, Earth and Planetary Astrophysics, Instrumentation and Methods for Astrophysics},\n")
+    f.write(
+        "keywords = {Solar system astronomy, Small Solar System bodies, Astronomy software, Astronomical simulations, 1529, 1469, 1855, 1857, Earth and Planetary Astrophysics, Instrumentation and Methods for Astrophysics},\n"
+    )
     f.write("year = 2025,\n")
     f.write("month = aug,\n")
     f.write("volume = {170},\n")
@@ -36,14 +40,17 @@ def cite_sorcha(f=sys.stdout):
     f.write("adsnote = {Provided by the SAO/NASA Astrophysics Data System}\n")
     f.write("}\n")
 
-
     f.write("\n\n")
 
     f.write("@ARTICLE{2025AJ....170...97H,\n")
-    f.write("author = {{Holman}, Matthew J. and {Bernardinelli}, Pedro H. and {Schwamb}, Megan E. and {Juri{\'c}}, Mario and {Oldag}, Drew and {West}, Maxine and {Napier}, Kevin J. and {Merritt}, Stephanie R. and {Fedorets}, Grigori and {Cornwall}, Samuel and {Kurlander}, Jacob A. and {Eggl}, Siegfried and {Kubica}, Jeremy and {Kiker}, Kathleen and {Murtagh}, Joseph and {Naidu}, Shantanu P. and {Chandler}, Colin Orion},\n")
+    f.write(
[…]
