### 2025-09-01 2a84f95
Update sorchaConfigs.py
update default downloads to new JPL files - fixes #1165
