### 2025-09-02 885608c
add links to data files
add links to data files  so they can be downloaded to run notebooks separately from dev install
