### 2025-09-02 0012761
doc and JOSS submission updates
doc and JOSS submission updates
