### 2025-09-02 2936537
Add default-matching drop_unlinked config to .ini files
