+++ src/sorcha/data/survey_setups/Rubin_full_footprint.ini
@@ -140,6 +140,8 @@ SSP_track_window = 15
 # For the LSST, 12pm Chile Standard Time is 4pm UTC.
 SSP_night_start_utc = 16.0
 
+# Remove detections of unlinked objects
+drop_unlinked=True
 
 [OUTPUT]
 
