+++ src/sorcha/data/demo/sorcha_config_demo.ini
@@ -150,6 +150,8 @@ SSP_track_window = 15
 # For the LSST, 12pm Chile Standard Time is 4pm UTC.
 SSP_night_start_utc = 16.0
 
+# Remove detections of unlinked objects
+drop_unlinked=True
 
 [OUTPUT]
 
