+++ docs/postprocessing.rst
@@ -246,7 +246,7 @@ Lightcurve Template Class
 Sinusoidal Light Curve  Class
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-Inside the `Sorcha add-ons  GitHub repository <https://github.com/dirac-institute/sorcha-addons>`_, we provide a simple example implementation where the apparent magnitude of the object (that is, the magnitude after all geometric effects have been taken into account), has a sinusoidal term added to it. To use this function, in the :ref:`CPP` file, the user must provide a light curve amplitude (`LCA`), corresponding to half the peak-to-peak amplitude for the magnitude changes, a period `Period`, and a reference time `Time0` where the light curve is at 0 - if these are not provided, the software will produce an error message. Despite being simple, that implementation shows all the class methods that need to be implemented for a custom light curve function. We have an `example Jupyter notebook <notebooks/demo_Lightcurve.ipynb>`_  demonstrating the SinusoidalLightCurve class built into `Sorcha add-ons package <https://github.com/dirac-institute/sorcha-addons>`_,  To use this prescription, the **lc_model** :ref:`configuration file<configs>` variable should be set to **sinusoidal**.
+Inside the `Sorcha add-ons  GitHub repository <https://github.com/dirac-institute/sorcha-addons>`_, we provide a simple example implementation where the apparent magnitude of the object (that is, the magnitude after all geometric effects have been taken into account), has a sinusoidal term added to it. To use this function, in the :ref:`CPP` file, the user must provide a light curve amplitude (`LCA`), corresponding to half the peak-to-peak amplitude for the magnitude changes, a period `Period`, and a reference time `Time0` where the light curve is at 0 - if these are not provided, the software will produce an error message. Despite being simple, that implementation shows all the class methods that need to be implemented for a custom light curve function. We have an `example Jupyter notebook <notebooks/demo_Lightcurve.ipynb>`__  demonstrating the SinusoidalLightCurve class built into `Sorcha add-ons package <https://github.com/dirac-institute/sorcha-addons>`_,  To use this prescription, the **lc_model** :ref:`configuration file<configs>` variable should be set to **sinusoidal**.
 
 .. _vignetting:
 
@@ -281,7 +281,7 @@ further from the center of the FOV have shallower depths.
   ``Sorcha`` currently only has  a vignetting model for the LSSTCam.
 
 .. seealso::
-  We have a `Jupyter notebook <notebooks/demo_Vignetting.ipynb>`_  demonstrating ``Sorcha``'s vignetting calculation. 
+  We have a `Jupyter notebook <notebooks/demo_Vignetting.ipynb>`__  demonstrating ``Sorcha``'s vignetting calculation. 
 
 .. _footprint:
 
@@ -325,7 +325,7 @@ To include this filter, the following options should be set in the :ref:`configs
 
 
 .. seealso::
-  We have a `Jupyter notebook <notebooks/demo_CircleFootprint.ipynb>`_  demonstrating ``Sorcha``'s circle radius (simple sensor area) filter.
+  We have a `Jupyter notebook <notebooks/demo_CircleFootprint.ipynb>`__  demonstrating ``Sorcha``'s circle radius (simple sensor area) filter.
 
 .. _full_camera_footprint:
 
@@ -365,7 +365,7 @@ Additionally, the camera footprint model can account for the losses at the edge 
 
 
 .. seealso::
-  We have a `Jupyter notebook <notebooks/demo_FootprintFilter.ipynb>`_  demonstrating ``Sorcha``'s full camera footprint filter.
[…]
