### 2025-09-02 92118a7
updates to docs and JOSS submission
- updates to docs and JOSS submission

- also removing temp file that should not have been added to the last push to this PR
