+++ JOSS_submission/paper.bib
@@ -465,7 +465,7 @@ archivePrefix = {arXiv},
 
 
 @ARTICLE{rein2012,
-       author = {{Rein}, H. and {Liu}, S. -F.},
+       author = {Rein, H. and Liu, S.-F.},
         title = "{REBOUND: an open-source multi-purpose N-body code for collisional dynamics}",
       journal = {A\&A},
      keywords = {methods: numerical, planets and satellites: rings, protoplanetary disks, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Mathematics - Dynamical Systems, Physics - Computational Physics},
@@ -484,7 +484,7 @@ archivePrefix = {arXiv},
 
 
 @ARTICLE{rein2015,
-       author = {{Rein}, Hanno and {Spiegel}, David S.},
+       author = {Rein, H. and Spiegel, D. S.},
         title = "{IAS15: a fast, adaptive, high-order integrator for gravitational dynamics, accurate to machine precision over a billion orbits}",
       journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {gravitation, methods: numerical, planets and satellites: dynamical evolution and stability, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics, Mathematics - Numerical Analysis},
