+++ sorcha_citation.txt
@@ -1,603 +1,0 @@
-
-Sorcha: 
-
-@ARTICLE{2025AJ....170..100M,
-author = {{Merritt}, Stephanie R. and {Fedorets}, Grigori and {Schwamb}, Megan E. and {Cornwall}, Samuel and {Bernardinelli}, Pedro H. and {Juri{'c}}, Mario and {Holman}, Matthew J. and {Kurlander}, Jacob A. and {Eggl}, Siegfried and {Oldag}, Drew and {West}, Maxine and {Kubica}, Jeremy and {Murtagh}, Joseph and {Jones}, R. Lynne and {Yoachim}, Peter and {Lyttle}, Ryan R. and {Kelley}, Michael S.~P. and {Moeyens}, Joachim and {Kiker}, Kathleen and {Naidu}, Shantanu P. and {Snodgrass}, Colin and {Matthews}, Shannon M. and {Chandler}, Colin Orion},
-title = "{Sorcha: A Solar System Survey Simulator for the Legacy Survey of Space and Time}"
-,journal = {j},
-keywords = {Solar system astronomy, Small Solar System bodies, Astronomy software, Astronomical simulations, 1529, 1469, 1855, 1857, Earth and Planetary Astrophysics, Instrumentation and Methods for Astrophysics},
-year = 2025,
-month = aug,
-volume = {170},
-number = {2},
-eid = {100},
-pages = {100},
-doi = {10.3847/1538-3881/add3ec},
-archivePrefix = {arXiv},
-eprint = {2506.02804},
-primaryClass = {astro-ph.EP},
-adsurl = {https://ui.adsabs.harvard.edu/abs/2025AJ....170..100M},
-adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-
-@ARTICLE{2025AJ....170...97H,
-author = {{Holman}, Matthew J. and {Bernardinelli}, Pedro H. and {Schwamb}, Megan E. and {Juri{'c}}, Mario and {Oldag}, Drew and {West}, Maxine and {Napier}, Kevin J. and {Merritt}, Stephanie R. and {Fedorets}, Grigori and {Cornwall}, Samuel and {Kurlander}, Jacob A. and {Eggl}, Siegfried and {Kubica}, Jeremy and {Kiker}, Kathleen and {Murtagh}, Joseph and {Naidu}, Shantanu P. and {Chandler}, Colin Orion},
-title = "{Sorcha: Optimized Solar System Ephemeris Generation}",
-journal = {j},
-keywords = {Ephemerides, Small Solar System bodies, N-body simulations, Sky surveys, 464, 1469, 1083, 1464, Earth and Planetary Astrophysics, Instrumentation and Methods for Astrophysics, Computational Physics},
-year = 2025,
-month = aug,
-volume = {170},
-number = {2},
-eid = {97},
-pages = {97},
-doi = {10.3847/1538-3881/ade435},
-archivePrefix = {arXiv},
-eprint = {2506.02140},
-primaryClass = {astro-ph.EP},
- adsurl = {https://ui.adsabs.harvard.edu/abs/2025AJ....170...97H},
- adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-
-Please also cite the packages that are utilized by Sorcha:
-
-ASSIST:
-
-@ARTICLE{2023PSJ.....4...69H,
-author = {{Holman}, Matthew J. and {Akmal}, Arya and {Farnocchia}, Davide and {Rein}, Hanno and {Payne}, Matthew J. and {Weryk}, Robert and {Tamayo}, Daniel and {Hernandez}, David M.},
[…]
