+++ JOSS_submission/paper.md
@@ -120,7 +120,7 @@ aas-journal: Astronomical Journal
 
 # Statement of Need
 
-The upcoming Legacy Survey of Space and Time (LSST) at the Vera C. Rubin Observatory [@lsstsciencebook2009; @ivezic2019; @bianco2022] is expected to revolutionize solar system astronomy. Unprecedented in scale, this ten-year wide-field survey will take ~2 million exposures split between 6 filters while also discovering and monitoring millions more solar system objects than are currently known [@jones2009; @jones2018; @lsstsciencebook2009; @solontoi2010; @shannon2015; @grav2016; @silsbee2016; @veres2017; @schwamb2018; @ivezic2019; @fedorets2020; @hoover2022; @kurlander2025; @murtagh2025]. This wealth of new information surpasses any survey to date in its combination of depth, sky coverage and sheer number of observations, The LSST will enable planetary astronomers to probe the dynamics and formation history of the solar system on a scale never before attempted. However, all astronomical surveys are affected by a complex set of intertwined observational biases, including observational strategy and cadence, limiting magnitude, instrumentation effects and observing conditions. The small body discoveries from an astronomical survey therefore provide a biased and distorted view of the actual underlying population. To help address this, survey simulators have emerged as powerful tools for assessing the impact of observational biases and aiding in the study of the target population. Open source survey simulators have long been used in smaller population-specific surveys such as the Canada–France Ecliptic Plane Survey (CFEPS) [@jones2006; @kavelaars2009; @petit2011] and the Outer Solar System Origins Survey (OSSOS) [@bannister2016; @bannister2018; @lawler2018] to forward-model the effects of biases on a given population, allowing for a direct comparison to real discoveries. However, there is no commonly-used package available already suitable for the job for future petatybe Solar System discovery surveys.  The CFEPS and OSSOS survey simulators are specifically designed for their bespoke surveys. The scale and tremendous scope of the LSST requires the development of a new tool capable of handling the scale of the Rubin Observatory’s LSST and all solar system small body populations.
[…]
