### 2025-09-09 #8145 @Onoddil
That all looks good to me now, thanks for the detailed response! I don't know if these can be caught by the JOSS team in the "accept the paper" stage, but the new sentence in the paper re: no comparative software just needs a quick copyedit "...for the job **ofhandling** future **petatybe**...". I figure this can just be sorted in post since it's a very dull fixup, so I've gone ahead and ticked all of the options in my review checklist and am happy to sign off on the paper for acceptance and publication.
