### 2025-09-17 5a220c9
bugged foxed edge case
bug fixed edge case with Visits footprint filter and added unit test to make sure this is covered.

removed linking filter from DES. now it just uses the Discovery Filter with a new flag turning it on and off

updated and added unit tests
