### 2025-09-17 77b7996
Initial editorial tweaks for JOSS paper
