+++ JOSS_submission/paper.bib
@@ -23,7 +23,7 @@
     pages = {9-12},
     doi = {10.1016/j.pss.2017.02.013},
     adsurl = {https://ui.adsabs.harvard.edu/abs/2018P&SS..150....9A},
-    adsnote = {Provided by the SAO/NASA Astrophysics Data System} 
+    adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @article{astropy2013,
@@ -39,7 +39,7 @@
     Month = oct,
     Pages = {A33},
     Primaryclass = {astro-ph.IM},
-    Title = {{Astropy: A community Python package for astronomy}},
+    Title = {Astropy: A community {Python} package for astronomy},
     Volume = 558,
     Year = 2013,
     Bdsk-Url-1 = {https://dx.doi.org/10.1051/0004-6361/201322068}
@@ -89,8 +89,8 @@
          {van Kerkwijk}, M.~H. and {de la Vega}, A. and {Watkins}, L.~L. and
          {Weaver}, B.~A. and {Whitmore}, J.~B. and {Woillez}, J. and
          {Zabalza}, V. and {Astropy Contributors}},
-        title = "{The Astropy Project: Building an Open-science Project and Status of the v2.0 Core Package}",
-      journal = {Astronomical Journal},
+        title = {The {Astropy} {Project}: Building an Open-science Project and Status of the v2.0 Core Package},
+      journal = {The Astronomical Journal},
      keywords = {methods: data analysis, methods: miscellaneous, methods: statistical, reference systems, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2018,
         month = sep,
@@ -108,8 +108,8 @@ archivePrefix = {arXiv},
 
 @ARTICLE{astropy2022,
[…]
