### 2025-09-17 2ba1e65
Add DOI for Lawler et al. (2018)
