+++ JOSS_submission/paper.bib
@@ -403,6 +403,7 @@ archivePrefix = {arXiv},
     PAGES={14},
     YEAR={2018},
     URL={https://www.frontiersin.org/article/10.3389/fspas.2018.00014},
+    DOI={10.3389/fspas.2018.00014},
     ISSN={2296-987X}
 }
 
