+++ src/sorcha/DP1.py
@@ -1,351 +1,0 @@
-#!/usr/bin/python
-
-import sys
-import time
-import numpy as np
-import argparse
-import os
-import logging
-
-from sorcha.ephemeris.simulation_driver import create_ephemeris
-from sorcha.ephemeris.simulation_setup import precompute_pointing_information
-
-from sorcha.modules.PPReadPointingDatabase import PPReadPointingDatabase
-from sorcha.modules.PPLinkingFilter import PPLinkingFilter
-from sorcha.modules.PPTrailingLoss import PPTrailingLoss
-from sorcha.modules.PPBrightLimit import PPBrightLimit
-from sorcha.modules.PPCalculateApparentMagnitude import PPCalculateApparentMagnitude
-from sorcha.modules.PPApplyFOVFilter import PPApplyFOVFilter
-from sorcha.modules.PPSNRLimit import PPSNRLimit
-from sorcha.modules import PPAddUncertainties, PPRandomizeMeasurements
-from sorcha.modules import PPVignetting
-from sorcha.modules.PPFadingFunctionFilter import PPFadingFunctionFilter
-from sorcha.modules.PPFaintObjectCullingFilter import PPFaintObjectCullingFilter
-
-
-from sorcha.modules.PPMatchPointingToObservations import PPMatchPointingToObservations
-from sorcha.modules.PPMagnitudeLimit import PPMagnitudeLimit
-from sorcha.modules.PPOutput import PPWriteOutput, PPIndexSQLDatabase
-from sorcha.modules.PPGetMainFilterAndColourOffsets import PPGetMainFilterAndColourOffsets
-from sorcha.modules.PPFootprintFilter import Footprint
-from sorcha.modules.PPStats import stats
-
-from sorcha.readers.CombinedDataReader import CombinedDataReader
-from sorcha.readers.CSVReader import CSVDataReader
-from sorcha.readers.EphemerisReader import EphemerisDataReader
-from sorcha.readers.OrbitAuxReader import OrbitAuxReader
-
-from sorcha.activity.activity_registration import update_activity_subclasses
-from sorcha.lightcurves.lightcurve_registration import update_lc_subclasses
-
-from sorcha.utilities.sorchaArguments import sorchaArguments
-from sorcha.utilities.sorchaConfigs import sorchaConfigs, PrintConfigsToLog
-from sorcha.utilities.sorchaCommandLineParser import sorchaCommandLineParser
-from sorcha.utilities.fileAccessUtils import FindFileOrExit
-from sorcha.utilities.citation_text import cite_sorcha
-from sorcha.utilities.sorchaGetLogger import sorchaGetLogger
-
-
-def runDP1Simulation(args, sconfigs):
-    """
-    Runs the post processing survey simulator functions that apply a series of
-    filters to bias a model Solar System small body population to what the
-    Vera C. Rubin Observatory Legacy Survey of Space and Time would observe.
-
-    Parameters
-    -----------
-    args : dictionary or `sorchaArguments` object
-        dictionary of command-line arguments.
-
-    pplogger : logging.Logger, optional
-        The logger to use in this function. If None creates a new one.
-        Default = None
-    sconfigs: dataclass
-        Dataclass of configuration file arguments.
-
-    Returns
-    -----------
-    None.
-
-    """
-    pplogger = logging.getLogger(__name__)
-    pplogger.info("Post-processing begun.")
-
-    try:
-        args.validate_arguments()
-    except Exception as err:
-        pplogger.error(err)
-        sys.exit(err)
-
-    # if verbosity flagged, the verboselog function will log the message specified
-    # if not, verboselog does absolutely nothing
[…]
