### 2025-09-18 1fa0c8f
Update PPVisitsFootprintFilter.py
