+++ src/sorcha/data/survey_setups/DES_config_file.ini
@@ -99,45 +99,16 @@ fading_function_on = True
 
 
 [LINKINGFILTER]
-; # For DES a single detection was considered a tracklet detection and 
-; # a minimum of 7 unique nights of detection are needed to classify an object over the entire survey time.
-; # The default Sorcha LinkingFilter is used to filter these detections and 
-; # then a Discovery Filter is used for more specific requirements for discovering objects with DES. see Bernardinelli et al (2022)
+# For DES a single detection was considered a tracklet detection and 
+# a minimum of 7 unique nights of detection are needed to classify an object over the entire survey time.
+# a Discovery Filter is used for more specific requirements for discovering objects with DES. see Bernardinelli et al (2022)
+des_discovery_on = True
 
-
-# # SSP detection efficiency. Which fraction of the objects will
-# # the automated Rubin Solar System Processing (SSP) pipeline successfully link? Float.
-SSP_detection_efficiency = 1
-
-
-# Length of tracklets. How many observations of an object during one night are
-# required to produce a valid tracklet?
-SSP_number_observations = 1
-
-# Minimum separation (in arcsec) between two observations of an object required 
-# for the linking software to distinguish them as separate and therefore as a valid tracklet.
-SSP_separation_threshold = 0.1
-
-# Maximum time separation (in days) between subsequent observations in a tracklet. 
-# Default is 0.0625 days (90mins).
-SSP_maximum_time = 0
-
-# Number of tracklets for detection. How many tracklets are required to classify
-# an object as detected? 
-SSP_number_tracklets = 7
-
-# The number of tracklets defined above must occur in <= this number of days to 
-# constitute a complete track/detection.
-SSP_track_window = 2192
-
-# The time in UTC at which it is noon at the observatory location (in standard time).
-# For the LSST, 12pm Chile Standard Time is 4pm UTC.
-SSP_night_start_utc = 16.0
 
 # distance (au) and motion (deg/day) cuts for detecting objects in DES:
 # Bernardinelli et al (2022) used 29 < d < 2500 au, for the detection of TNOs with DES
-distance_cut_upper = 2500
-distance_cut_lower = 29
+; distance_cut_upper = 2500
+; distance_cut_lower = 29
 
 ; motion_cut_upper = 
 ; motion_cut_lower = 
