### 2025-10-02 #8145 @warrickball
Hi @mschwamb, I just wanted to check in on your progress here. I apologise if I wasn't clear in [my previous comment](https://github.com/openjournals/joss-reviews/issues/8145#issuecomment-3273709569) but at the moment the publication process is waiting for you to

> * make a new tagged release and let me know the version number (I guess this will be the PyPI/Conda version);
> * archive the reviewed version on a service like Zenodo, Figshare or similar institutional repository;
> * double check that the metadata  (title, licence, authors) matches the JOSS paper; and
> * let me know the DOI of the deposit here.

Let me know if you have any questions about the process.
