### 2025-10-03 3723903
DES bug fixes
Added bug fixes for edge cases and unit tests
