+++ src/sorcha/modules/PPVisitsFootprintFilter.py
@@ -17,6 +17,8 @@ def PPVisitsFootprint(
     ra_name="RA_deg",
     dec_name="Dec_deg",
     fieldId="FieldID",
+    k_num=3,
+    plot=False,
 ):
     """
     Determine whether detections fall on the sensors defined by the
@@ -35,6 +37,8 @@ def PPVisitsFootprint(
         Column names in field_df for RA, Dec, and field ID respectively.
     ephermers_buffer : float
         accounts for ar wraping RA points around 0-360, while yhe camera footprint doesnt to avoid the need to create complex shapes
+    k_num : int
+        number of ccds to check
     Returns
     -------
     detected_list : list
@@ -80,15 +84,47 @@ def PPVisitsFootprint(
                     "Ephemides buffer is too big to account for wrap around. Your footprint will be inaccurate"
                 )
                 # this might casue issues with objects of massive ephermides (greater than 180). Maybe warn the user of the super extreme case?
+            # ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
+            # Prepare combined list of centers and corners
+            all_ccd_points = []
+            point_to_ccd_map = []  # Maps each point index to its CCD index
 
-            ccd_centers = [(row["ra_centre"], row["dec_centre"]) for row in rows]
-            ccd_tree = KDTree(ccd_centers)
-            if len(ccd_centers) > 1:
-                k = min(3, len(ccd_centers))  # making sure theirs no error with 2 ccds
-                _, closest_ccd_indices = ccd_tree.query(points_query, k=k)
-                unique_ccd_indices = set(closest_ccd_indices.flatten())  # remove duplicates
+            for idx, row in enumerate(rows):
+                # Add center
+                all_ccd_points.append((row["ra_centre"], row["dec_centre"]))
+                point_to_ccd_map.append(idx)
+
+                # Add corners
+                corners = [
+                    (row["llcra"], row["llcdec"]),
+                    (row["lrcra"], row["lrcdec"]),
+                    (row["urcra"], row["urcdec"]),
+                    (row["ulcra"], row["ulcdec"]),
+                ]
+                all_ccd_points.extend(corners)
+                point_to_ccd_map.extend([idx] * 4)
+
+            # Build KDTree with all points
+            ccd_tree = KDTree(all_ccd_points)
+
+            if len(rows) > 1:
+                k = min(k_num, len(all_ccd_points))
+                _, closest_point_indices = ccd_tree.query(points_query, k=k)
+
+                # Map back to unique CCD indices
+                unique_ccd_indices = set(point_to_ccd_map[i] for i in closest_point_indices.flatten())
             else:
-                unique_ccd_indices = [0]
+                unique_ccd_indices = {0}
+
+            # ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
+            # ccd_centers = [(row["ra_centre"], row["dec_centre"]) for row in rows]
+            # ccd_tree = KDTree(ccd_centers)
+            # if len(ccd_centers) > 1:
+            #     k = min(k_num, len(ccd_centers))  # making sure theirs no error with 2 ccds
+            #     _, closest_ccd_indices = ccd_tree.query(points_query, k=k)
+            #     unique_ccd_indices = set(closest_ccd_indices.flatten())  # remove duplicates
+            # else:
+            #     unique_ccd_indices = [0]
             # create polygons for closest ccds to points and record detector name/id
             polygons = []
             detectors = []
@@ -119,6 +155,61 @@ def PPVisitsFootprint(
                         detector_for_index[idx_in_df] = detectors[poly_idx]
                         lim_mag_list[idx_in_df] = limmag[poly_idx]
                         break  # no need to check other polygons for this point if already on one
+            # ——————————————————————————————— plotting ———————————————————————————————————
+            if plot == True:
[…]
