### 2025-10-05 #8145 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
