### 2025-10-05 bbf3ef8
Update paper.md
add USA to LINCC Frameworks address
