### 2025-10-09 ac729a6
changes and fixes to VisitsFootprint
