+++ src/sorcha/modules/PPVisitsFootprintFilter.py
@@ -39,6 +39,8 @@ def PPVisitsFootprint(
         accounts for ar wraping RA points around 0-360, while yhe camera footprint doesnt to avoid the need to create complex shapes
     k_num : int
         number of ccds to check
+    plot : boolen
+        used to turning on plotting for debugging
     Returns
     -------
     detected_list : list
@@ -84,26 +86,25 @@ def PPVisitsFootprint(
                     "Ephemides buffer is too big to account for wrap around. Your footprint will be inaccurate"
                 )
                 # this might casue issues with objects of massive ephermides (greater than 180). Maybe warn the user of the super extreme case?
-            # ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
-            # Prepare combined list of centers and corners
-            all_ccd_points = []
-            point_to_ccd_map = []  # Maps each point index to its CCD index
 
+
+            # Kd tree using corners and centres of ccds
+            
+            n = len(rows)
+            # make an empty array to contain all ccd ra and decs
+            all_ccd_points = np.empty((n * 5, 2), dtype=float)
+            # make a numpy array for indexing
+            point_to_ccd_map = np.repeat(np.arange(n), 5)
+
+            # create the numpy array
             for idx, row in enumerate(rows):
-                # Add center
-                all_ccd_points.append((row["ra_centre"], row["dec_centre"]))
-                point_to_ccd_map.append(idx)
-
-                # Add corners
-                corners = [
-                    (row["llcra"], row["llcdec"]),
-                    (row["lrcra"], row["lrcdec"]),
-                    (row["urcra"], row["urcdec"]),
-                    (row["ulcra"], row["ulcdec"]),
+                all_ccd_points[idx*5:(idx+1)*5] = [
+                    (row["ra_centre"], row["dec_centre"]),   
+                    (row["llcra"], row["llcdec"]),         
+                    (row["lrcra"], row["lrcdec"]),         
+                    (row["urcra"], row["urcdec"]),         
+                    (row["ulcra"], row["ulcdec"]),         
                 ]
-                all_ccd_points.extend(corners)
-                point_to_ccd_map.extend([idx] * 4)
-
             # Build KDTree with all points
             ccd_tree = KDTree(all_ccd_points)
 
@@ -111,24 +112,15 @@ def PPVisitsFootprint(
                 k = min(k_num, len(all_ccd_points))
                 _, closest_point_indices = ccd_tree.query(points_query, k=k)
 
-                # Map back to unique CCD indices
+                # Map back to unique ccd indices
                 unique_ccd_indices = set(point_to_ccd_map[i] for i in closest_point_indices.flatten())
             else:
                 unique_ccd_indices = {0}
 
-            # ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
-            # ccd_centers = [(row["ra_centre"], row["dec_centre"]) for row in rows]
-            # ccd_tree = KDTree(ccd_centers)
-            # if len(ccd_centers) > 1:
-            #     k = min(k_num, len(ccd_centers))  # making sure theirs no error with 2 ccds
-            #     _, closest_ccd_indices = ccd_tree.query(points_query, k=k)
-            #     unique_ccd_indices = set(closest_ccd_indices.flatten())  # remove duplicates
-            # else:
-            #     unique_ccd_indices = [0]
-            # create polygons for closest ccds to points and record detector name/id
             polygons = []
             detectors = []
             limmag = []
+            # create polygons of only ccds needed
             for idx in unique_ccd_indices:
                 row = rows[idx]
[…]
