### 2025-10-09 97f292c
fixed bug with Discoveryfilter
