### 2025-10-11 e38c01a
fix left sidebar and other tweaks
