### 2025-10-13 3ccf452
added doc strings and cleaned up parts of code
