+++ src/sorcha/modules/PPVisitsFootprintFilter.py
@@ -151,58 +151,7 @@ def PPVisitsFootprint(
             
             # ——————————————————————————————— plotting ———————————————————————————————————
             if plot == True:
-                # Plotting code for each camera footprint use only for debugging (handy for unit tests)
-                plt.figure(figsize=(8, 8))
-                # Plot ccds with created polygons
-                for idx, poly in enumerate(polygons):
-                    x, y = poly.exterior.xy
-                    if idx == 0:
-                        plt.plot(x, y, "b-", linewidth=2, label="Used CCDs")
-                    else:
-                        plt.plot(x, y, "b-", linewidth=2)
-
-                polygons_plot = []
-                # plot unused ccds for checks
-                for j, row in enumerate(rows):
-                    if j not in unique_ccd_indices:
-                        corners = np.array(
-                            [
-                                (row["llcra"], row["llcdec"]),
-                                (row["lrcra"], row["lrcdec"]),
-                                (row["urcra"], row["urcdec"]),
-                                (row["ulcra"], row["ulcdec"]),
-                                (row["llcra"], row["llcdec"]),  # closing polygon
-                            ]
-                        )
-                        polygons_plot.append(Polygon(corners))
-                    for poly in polygons_plot:
-                        x, y = poly.exterior.xy
-
-                        if j == 0:
-                            plt.plot(x, y, "r-", linewidth=2, label="Not Used CCDs")
-                        else:
-                            plt.plot(x, y, "r-", linewidth=2)
-                # Plot points
-                for row in points_query:
-
-                    ra = row[0]
-
-                    dec = row[1]
-
-                    if any(poly.contains(Point(ra, dec)) for poly in polygons):
-                        plt.scatter(ra, dec, c="g", marker="o", label="Detected")
-                    else:
-                        plt.scatter(ra, dec, c="r", marker="o", label="Not Detected")
-
-                handles, labels = plt.gca().get_legend_handles_labels()
-                unique = dict(zip(labels, handles))
-                plt.legend(unique.values(), unique.keys())
-                plt.title(f"Camera Footprint and Detections for obs_id {obs_id}")
-                plt.xlabel("RA (deg)")
-                plt.ylabel("Dec (deg)")
-                plt.tight_layout()
-                plt.grid(True)
-                plt.show()
+                _plotting_camera_footprint(polygons,rows,unique_ccd_indices,points_query,obs_id)
             # ——————————————————————————————— end of plotting ———————————————————————————————————
 
     detected_list = list(detected_indices)  # list of detected observations
@@ -215,175 +164,10 @@ def PPVisitsFootprint(
 
 
 
-
-
-
-
-
-import logging
-import sqlite3
-from collections import defaultdict
-
-import numpy as np
-import matplotlib.pyplot as plt
-from matplotlib.path import Path
-from scipy.spatial import KDTree
-
-logger = logging.getLogger(__name__)
-
-# @profile
-def _PPVisitsFootprint(
-    field_df,
-    query,
-    visits_filename,
-    ephermers_buffer=1.5,
-    ra_name="RA_deg",
-    dec_name="Dec_deg",
-    fieldId="FieldID",
-    k_num=3,
-    plot=False,
-):
-    """
-    Determine whether detections fall on the sensors defined by the
-    footprint. Also returns the ID for the sensor a detection is made
-    on.
-
-    Parameters
-    ----------
-    field_df : pandas.DataFrame
-        DataFrame containing detection information with pointings.
-    query : str
[…]
