+++ src/sorcha/modules/DESDiscoveryFilter.py
@@ -89,7 +89,7 @@ def DESDiscoveryFilter(
 def compute_arccut(times):
     """
     Computes ARCCUT, the time between the first and last detection dropping one night of detection
-    
+
     Parameters
     -----------
     times: list
@@ -97,7 +97,7 @@ def compute_arccut(times):
 
     Returns
     ----------
-    flag: bool 
+    flag: bool
         returns true if ARCCUT is greater than 6 months
     """
     arccut = 0.0
@@ -131,12 +131,12 @@ def compute_triplet(times, thresh):
     -----------
     times: list
         list of times, must be in DAYS
-    thresh: 
+    thresh:
         threshold time for pairs
 
     Returns
     ----------
-    flag: bool 
+    flag: bool
         returns true if triplet is formed
 
     """
