### 2025-10-13 0d78d90
renamed parameter k_num to ccd_num
renamed k_num to ccd_num in Visits footprint
