+++ JOSS_submission/paper.md
@@ -24,7 +24,7 @@ authors:
   - name: Mario Jurić
     orcid:  0000-0003-1996-9252
     affiliation: "5"
-  - name: Matthew J.Holman
+  - name: Matthew J. Holman
     orcid:  0000-0002-1139-4880
     affiliation: "6"
   - name: Jacob A. Kurlander
