### 2025-10-13 28ea37c
realign left sidebar now that author list is wrapped differently
