### 2025-10-17 ed96949
Update paper.md
update corresponding author
