+++ JOSS_submission/paper.md
@@ -8,13 +8,13 @@ authors:
   - name: Stephanie R. Merritt
     orcid: 0000-0001-5930-2829
     affiliation: "1"
-    corresponding: true
   - name: Grigori Fedorets
     orcid: 0000-0002-8418-4809
     affiliation: "1,2,3"
   - name: Megan E. Schwamb
     orcid: 0000-0003-4365-1455
     affiliation: "1"
+    corresponding: true
   - name: Samuel Cornwall
     orcid: 0000-0002-0672-5104
     affiliation: "4"
