### 2025-10-17 #8145 @warrickball
Congratulations @mschwamb, your paper has been published in JOSS! Thanks again for your patience, especially with fixing up our pipeline at the very end.

Thanks also to @Onoddil & @Professor-G for reviewing! JOSS simply wouldn't be possible without its community of volunteers.
